remove unused services
