--- src/UI/Components/Diagram/Mocks/Mock.ts
@@ -1,4706 +1,0 @@
-import { ServiceInstanceModel, ServiceModel } from "@/Core";
-import { InstanceWithRelations } from "@/Data/Managers/V2/GETTERS/GetInstanceWithRelations";
-import { ComposerServiceOrderItem } from "@/UI/Components/Diagram/interfaces";
-
-export const testInstance: ServiceInstanceModel = {
-  id: "d938d5bb-8bf1-4b41-9e17-ae0b5069cbbf",
-  environment: "ff6dd80a-e18b-4a06-85fc-9a5cf4b45d0b",
-  service_entity: "mpn",
-  version: 4,
-  config: {},
-  state: "up",
-  candidate_attributes: null,
-  active_attributes: {
-    mcc: "001",
-    mnc: "01",
-    name: "bart",
-    site: "mpn_site",
-    tenant: "mpn",
-    peerings: [
-      {
-        vlan: 20,
-        routes: ["10.0.20.128/32,10.0.20.1", "10.0.20.144/32,10.0.20.1"],
-        switch: "switch-1",
-        peer_asn: 64999,
-        port_name: "xe-0/0/7",
-        peer_address: "10.0.20.1",
-        local_address: "10.0.20.0/31",
-      },
-    ],
-    mgmt_prefix: "172.26.100.0/24",
-    tenant_group: "mpn_group",
-  },
-  rollback_attributes: null,
-  created_at: "2023-05-16T07:40:28.868598",
-  last_updated: "2023-05-16T07:49:41.067413",
-  callback: [],
-  referenced_by: [],
-  deleted: false,
-  deployment_progress: {
-    total: 335,
-    failed: 0,
-    deployed: 335,
-    waiting: 0,
-  },
-  service_identity_attribute_value: "bart",
-};
-
-export const testService: ServiceModel = {
-  attributes: [
-    {
-      name: "name",
-      description: "The name of the mpn to deploy.",
-      modifier: "rw",
-      type: "string",
-      default_value: null,
-      default_value_set: false,
-      validation_type: null,
-      validation_parameters: null,
-    },
-    {
-      name: "mgmt_prefix",
-      description:
-        "The network to use for the management range. This needs to be at least a /25\nso that two /26 can be created from it.",
-      modifier: "rw",
-      type: "string",
-      default_value: null,
-      default_value_set: false,
-      validation_type: "ipaddress.IPv4Network",
-      validation_parameters: null,
-    },
-    {
-      name: "mcc",
-      description: "Mobile country code",
-      modifier: "rw",
-      type: "string",
-      default_value: null,
-      default_value_set: false,
-      validation_type: null,
-      validation_parameters: null,
-    },
-    {
-      name: "mnc",
-      description: "Mobile network code",
-      modifier: "rw",
-      type: "string",
-      default_value: null,
-      default_value_set: false,
-      validation_type: null,
-      validation_parameters: null,
-    },
-    {
-      name: "tenant_group",
-      description: "desc",
-      modifier: "rw",
-      type: "string",
-      default_value: null,
-      default_value_set: false,
-      validation_type: null,
-      validation_parameters: null,
-    },
-    {
-      name: "tenant",
-      description: "desc",
-      modifier: "rw",
-      type: "string",
-      default_value: null,
-      default_value_set: false,
-      validation_type: null,
-      validation_parameters: null,
-    },
-    {
-      name: "site",
-      description: "desc",
-      modifier: "rw",
-      type: "string",
-      default_value: null,
-      default_value_set: false,
-      validation_type: null,
-      validation_parameters: null,
-    },
-  ],
-  embedded_entities: [
-    {
-      name: "peerings",
-      description: "desc",
-      modifier: "rw+",
-      lower_limit: 0,
-      upper_limit: undefined,
-      attributes: [
-        {
-          name: "switch",
-          description: "The name/id of the switch to do the peering on",
-          modifier: "rw",
-          type: "string",
-          default_value: null,
-          default_value_set: false,
-          validation_type: "enum",
-          validation_parameters: {
-            names: {
-              "switch-1": "switch-1",
-              "switch-2": "switch-2",
-              "switch-3": "switch-3",
-              "switch-4": "switch-4",
-            },
-            value: "switch_t",
-          },
-        },
-        {
-          name: "port_name",
-          description:
-            "The name of the port on the switch to create the peering on",
-          modifier: "rw",
-          type: "string",
-          default_value: null,
-          default_value_set: false,
-          validation_type: null,
-          validation_parameters: null,
-        },
-        {
-          name: "local_address",
-          description: "The address to set on the interface on the MPN",
-          modifier: "rw",
-          type: "string",
-          default_value: null,
-          default_value_set: false,
-          validation_type: "ipaddress.IPv4Interface",
-          validation_parameters: null,
-        },
-        {
-          name: "peer_address",
-          description:
-            "The address of the peer. When BPG is used, this is also the peering ip",
-          modifier: "rw",
-          type: "string?",
-          default_value: null,
-          default_value_set: true,
-          validation_type: "ipaddress.IPv4Address?",
-          validation_parameters: null,
-        },
-        {
-          name: "vlan",
-          description:
-            "The vlan ID to use for this peering. When the vlan is 0 the traffic is untagged. If the vlan is already used\ninternally in the deployment, a validation error will be given.",
-          modifier: "rw",
-          type: "int",
-          default_value: null,
-          default_value_set: false,
-          validation_type: null,
-          validation_parameters: null,
-        },
-        {
-          name: "peer_asn",
-          description: "The ASN of the peer to setup the BGP session.",
-          modifier: "rw",
-          type: "int?",
-          default_value: null,
-          default_value_set: true,
-          validation_type: null,
-          validation_parameters: null,
-        },
-        {
-          name: "routes",
-          description:
-            "Routes are structured as prefix,gateway For example: 0.0.0.0/0,192.0.2.1",
-          modifier: "rw",
-          type: "string[]",
-          default_value: [] as string[],
-          default_value_set: true,
-          validation_type: "pydantic.constr[]",
-          validation_parameters: {
-            regex:
-              "^\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}\\/\\d{1,2},\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}$",
-          },
-        },
-      ],
-      embedded_entities: [],
-      inter_service_relations: [],
-      key_attributes: ["port_name", "switch", "vlan"],
-    },
-  ],
-  inter_service_relations: [],
-  environment: "ff6dd80a-e18b-4a06-85fc-9a5cf4b45d0b",
-  name: "mpn",
-  description:
-    "A redundant mobile private network deployment for Core and Transport",
-  lifecycle: {
-    states: [
-      {
-        name: "update_inprogress",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_start_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "failed",
-        label: "danger",
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "rejected",
-        label: "danger",
-        export_resources: false,
-        validate_self: "candidate",
-        validate_others: null,
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_acknowledged",
-        label: null,
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "up",
-        label: "success",
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_rejected",
-        label: "warning",
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "rollback",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "start",
-        label: null,
-        export_resources: false,
-        validate_self: "candidate",
-        validate_others: null,
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_acknowledged_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "deleting",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: true,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "creating",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_rejected_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_start",
-        label: null,
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "acknowledged",
-        label: null,
-        export_resources: false,
-        validate_self: "candidate",
-        validate_others: null,
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "terminated",
-        label: null,
-        export_resources: false,
-        validate_self: null,
-        validate_others: null,
-        purge_resources: false,
-        deleted: true,
-        values: {},
-      },
-    ],
-    transfers: [
-      {
-        source: "start",
-        target: "acknowledged",
-        error: "rejected",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: true,
-        config_name: null,
-        description: "initial validation",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "acknowledged",
-        target: "creating",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: false,
-        config_name: null,
-        description: "ack to creating",
-        target_operation: "promote",
-        error_operation: null,
-      },
-      {
-        source: "creating",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "creating done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "up",
-        target: "update_start",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "up to update_start",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_start",
-        target: "update_acknowledged",
-        error: "update_rejected",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: true,
-        config_name: null,
-        description: "update_start to update_ack",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_start_failed",
-        target: "update_acknowledged_failed",
-        error: "update_rejected_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: true,
-        config_name: null,
-        description: "update_start_failed to update_ack_failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_acknowledged",
-        target: "update_inprogress",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: false,
-        config_name: null,
-        description: "update_ack to update_inprogress",
-        target_operation: "promote",
-        error_operation: null,
-      },
-      {
-        source: "update_acknowledged_failed",
-        target: "update_inprogress",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: false,
-        config_name: null,
-        description: "update_ack_failed to update_inprogress_failed",
-        target_operation: "promote",
-        error_operation: null,
-      },
-      {
-        source: "update_inprogress",
-        target: "up",
-        error: "update_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_start",
-        target: "update_start",
-        error: "update_start_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update start failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected",
-        target: "update_rejected",
-        error: "update_rejected_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "upd_reject failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_start_failed",
-        target: "update_start",
-        error: "update_start_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update start recover",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected_failed",
-        target: "update_rejected",
-        error: "update_rejected_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "upd_reject recover",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected",
-        target: "up",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: true,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "abort update",
-        target_operation: "clear candidate",
-        error_operation: null,
-      },
-      {
-        source: "update_rejected_failed",
-        target: "failed",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: true,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "abort update failed",
-        target_operation: "clear candidate",
-        error_operation: null,
-      },
-      {
-        source: "update_failed",
-        target: "rollback",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: true,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rollback in update_failed",
-        target_operation: "rollback",
-        error_operation: null,
-      },
-      {
-        source: "rollback",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rollback done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "up",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "up to failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "failed",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "failed to up",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_failed",
-        target: "up",
-        error: "update_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update_failed to up",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "rejected",
-        target: "start",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rejected to start",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected",
-        target: "update_start",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update_rejected to start_update",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected_failed",
-        target: "update_start_failed",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update_rejected_failed to start_update_failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "deleting",
-        target: "terminated",
-        error: "deleting",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "deletion done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "rejected",
-        target: "terminated",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rejected to terminated",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "up",
-        target: "deleting",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "up to deleting",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "creating",
-        target: "deleting",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "creating to deleting",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_failed",
-        target: "deleting",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update_failed to deleting",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "failed",
-        target: "deleting",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "failed to deleting",
-        target_operation: null,
-        error_operation: null,
-      },
-    ],
-    initial_state: "start",
-    name: "lsm::simple",
-  },
-  config: {},
-  service_identity: "name",
-  service_identity_display_name: "MPN Name",
-  instance_summary: {
-    by_state: {
-      up: 1,
-      update_inprogress: 0,
-      update_start_failed: 0,
-      failed: 0,
-      rejected: 0,
-      update_failed: 0,
-      update_acknowledged: 0,
-      update_rejected: 0,
-      rollback: 0,
-      start: 0,
-      update_acknowledged_failed: 0,
-      deleting: 0,
-      creating: 0,
-      update_rejected_failed: 0,
-      update_start: 0,
-      acknowledged: 0,
-    },
-    by_label: {
-      no_label: 0,
-      info: 0,
-      success: 1,
-      danger: 0,
-      warning: 0,
-    },
-    total: 1,
-  },
-  strict_modifier_enforcement: true,
-  owner: null,
-  owned_entities: [],
-};
-
-export const testApiInstanceModel: ServiceModel = {
-  attributes: [
-    {
-      name: "name",
-      description: "",
-      modifier: "rw",
-      type: "string",
-      default_value: null,
-      default_value_set: false,
-      validation_type: null,
-      validation_parameters: null,
-    },
-    {
-      name: "service_id",
-      description: "",
-      modifier: "rw",
-      type: "string",
-      default_value: null,
-      default_value_set: false,
-      validation_type: null,
-      validation_parameters: null,
-    },
-    {
-      name: "should_deploy_fail",
-      description: "",
-      modifier: "rw+",
-      type: "bool",
-      default_value: false,
-      default_value_set: true,
-      validation_type: null,
-      validation_parameters: null,
-    },
-  ],
-  embedded_entities: [
-    {
-      name: "vlan_assigment_r1",
-      description: "",
-      modifier: "rw",
-      lower_limit: 1,
-      upper_limit: 1,
-      attributes: [
-        {
-          name: "address",
-          description:
-            "The IP-address/netmask to assign to the given VLAN interface.",
-          modifier: "rw+",
-
-          type: "string",
-          default_value: null,
-          default_value_set: false,
-          validation_parameters: null,
-        },
-        {
-          name: "vlan_id",
-          description: "The VLAN ID to assign to the given interface.",
-          modifier: "rw",
-
-          type: "int",
-          default_value: null,
-          default_value_set: false,
-          validation_type: "pydantic.conint",
-          validation_parameters: {
-            ge: 0,
-            lt: 4096,
-          },
-        },
-        {
-          name: "router_ip",
-          description:
-            "The IP address of the srlinux router that should be configured.",
-          modifier: "rw",
-
-          type: "string",
-          default_value: null,
-          default_value_set: false,
-          validation_parameters: null,
-        },
-        {
-          name: "interface_name",
-          description:
-            "The name of the interface that should be connected on the given VLAN.",
-          modifier: "rw",
-
-          type: "string",
-          default_value: null,
-          default_value_set: false,
-          validation_type: null,
-          validation_parameters: null,
-        },
-      ],
-      embedded_entities: [],
-      inter_service_relations: [],
-      key_attributes: ["interface_name", "router_ip", "vlan_id"],
-    },
-    {
-      name: "vlan_assigment_r2",
-      description: "",
-      modifier: "rw",
-      lower_limit: 1,
-      upper_limit: 1,
-      attributes: [
-        {
-          name: "address",
-          description:
-            "The IP-address/netmask to assign to the given VLAN interface.",
-          modifier: "rw+",
-
-          type: "string",
-          default_value: null,
-          default_value_set: false,
-
-          validation_parameters: null,
-        },
-        {
-          name: "vlan_id",
-          description: "The VLAN ID to assign to the given interface.",
-          modifier: "rw",
-          type: "int",
-          default_value: null,
-          default_value_set: false,
-          validation_type: "pydantic.conint",
-          validation_parameters: {
-            ge: 0,
-            lt: 4096,
-          },
-        },
-        {
-          name: "router_ip",
-          description:
-            "The IP address of the srlinux router that should be configured.",
-          modifier: "rw",
-          type: "string",
-          default_value: null,
-          default_value_set: false,
-          validation_parameters: null,
-        },
-        {
-          name: "interface_name",
-          description:
-            "The name of the interface that should be connected on the given VLAN.",
-          modifier: "rw",
-
-          type: "string",
-          default_value: null,
-          default_value_set: false,
-          validation_type: null,
-          validation_parameters: null,
-        },
-      ],
-      embedded_entities: [],
-      inter_service_relations: [],
-      key_attributes: ["interface_name", "router_ip", "vlan_id"],
-    },
-  ],
-  inter_service_relations: [],
-  environment: "89473b11-69dd-4b54-a9ed-eef74e680f80",
-  name: "embedded-entity-service",
-  description: "",
-  lifecycle: {
-    states: [
-      {
-        name: "update_inprogress",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "creating",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "terminated",
-        label: null,
-        export_resources: false,
-        validate_self: null,
-        validate_others: null,
-        purge_resources: false,
-        deleted: true,
-        values: {},
-      },
-      {
-        name: "setting_start",
-        label: null,
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "failed",
-        label: "danger",
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "start",
-        label: null,
-        export_resources: false,
-        validate_self: "candidate",
-        validate_others: null,
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "setting_inprogress",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "up",
-        label: "success",
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_start",
-        label: null,
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "rejected",
-        label: "danger",
-        export_resources: false,
-        validate_self: "candidate",
-        validate_others: null,
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "deleting",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: true,
-        deleted: false,
-        values: {},
-      },
-    ],
-    transfers: [
-      {
-        source: "start",
-        target: "creating",
-        error: "rejected",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: true,
-        config_name: null,
-        description: "initial validation",
-        target_operation: "promote",
-        error_operation: null,
-      },
-      {
-        source: "creating",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "creating done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "up",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "up to failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "failed",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "failed to up",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "rejected",
-        target: "start",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rejected to start",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "deleting",
-        target: "terminated",
-        error: "deleting",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "resources deleted",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "rejected",
-        target: "terminated",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rejected to terminated",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "up",
-        target: "deleting",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "up to deleting",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "creating",
-        target: "deleting",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "creating to deleting",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "failed",
-        target: "deleting",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "failed to deleting",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "up",
-        target: "update_start",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "up to update_start",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_start",
-        target: "update_inprogress",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: false,
-        config_name: null,
-        description: "update_start to update_inprogress",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_inprogress",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "up",
-        target: "setting_start",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: true,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "up to setting_start",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "setting_start",
-        target: "setting_inprogress",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: false,
-        config_name: null,
-        description: "setting_start to setting_inprogress ",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "setting_inprogress",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "setting done",
-        target_operation: null,
-        error_operation: null,
-      },
-    ],
-    initial_state: "start",
-    name: "frontend_model::simple-with-set-state",
-  },
-  config: {},
-  service_identity: "name",
-  service_identity_display_name: undefined,
-  strict_modifier_enforcement: true,
-  instance_summary: {
-    by_state: {
-      update_inprogress: 0,
-      creating: 0,
-      setting_start: 0,
-      failed: 0,
-      start: 0,
-      setting_inprogress: 0,
-      up: 0,
-      update_start: 0,
-      rejected: 0,
-      deleting: 0,
-    },
-    by_label: {
-      no_label: 0,
-      info: 0,
-      success: 0,
-      danger: 0,
-      warning: 0,
-    },
-    total: 0,
-  },
-  owner: null,
-  owned_entities: [],
-};
-
-export const testApiInstance: ComposerServiceOrderItem = {
-  instance_id: "ae6c9dd7-5392-4374-9f13-df3bb42bf0db",
-  service_entity: "embedded-entity-service",
-  config: {},
-  action: null,
-  attributes: {
-    name: "test-emb",
-    service_id: "ebd-123",
-    should_deploy_fail: false,
-  },
-};
-
-export const testEmbeddedApiInstances: ComposerServiceOrderItem[] = [
-  {
-    instance_id: "10591ae5-6840-4816-9851-6bee74afc2a5",
-    service_entity: "vlan_assigment_r1",
-    config: {},
-    action: null,
-    attributes: {
-      address: "1.2.3.5/32",
-      vlan_id: 1,
-      router_ip: "1.2.3.4",
-      interface_name: "eth0",
-    },
-    embeddedTo: "ae6c9dd7-5392-4374-9f13-df3bb42bf0db",
-  },
-  {
-    instance_id: "c0293b05-04d1-4fb4-93f1-f60d93d9739c",
-    service_entity: "vlan_assigment_r2",
-    config: {},
-    action: null,
-    attributes: {
-      address: "1.2.3.3/32",
-      vlan_id: 123,
-      router_ip: "1.2.3.1",
-      interface_name: "eth12",
-    },
-    embeddedTo: "ae6c9dd7-5392-4374-9f13-df3bb42bf0db",
-  },
-];
-
-export const testParentService: ComposerServiceOrderItem = {
-  instance_id: "6af44f75-ba4b-4fba-9186-cc61c3c9463c",
-  service_entity: "parent-service",
-  config: {},
-  action: null,
-  attributes: {
-    name: "test",
-    service_id: "test123",
-    should_deploy_fail: false,
-  },
-};
-
-export const interServiceRelations: ComposerServiceOrderItem[] = [
-  {
-    instance_id: "13920268-cce0-4491-93b5-11316aa2fc37",
-    service_entity: "child-service",
-    config: {},
-    action: "create",
-    attributes: {
-      name: "test123456789",
-      service_id: "123test",
-      should_deploy_fail: false,
-    },
-    relatedTo: new Map().set(
-      "6af44f75-ba4b-4fba-9186-cc61c3c9463c",
-      "parent_entity",
-    ),
-  },
-  {
-    instance_id: "a4218978-c9ad-4fd8-95e4-b9e9a8c3c653",
-    service_entity: "container-service",
-    config: {},
-    action: "create",
-    attributes: {
-      name: "test12345",
-      service_id: "test12345",
-      should_deploy_fail: false,
-    },
-  },
-  {
-    instance_id: "0ede01ac-25c0-4029-9c7e-62ec6f1dcd54",
-    service_entity: "child_container",
-    config: {},
-    action: "create",
-    attributes: {
-      name: "child123",
-    },
-    embeddedTo: "a4218978-c9ad-4fd8-95e4-b9e9a8c3c653",
-    relatedTo: new Map().set(
-      "6af44f75-ba4b-4fba-9186-cc61c3c9463c",
-      "parent_entity",
-    ),
-  },
-];
-
-export const mockedInstanceWithRelations: InstanceWithRelations = {
-  instance: {
-    id: "085cdf92-0894-4b82-8d46-1dd9552e7ba3",
-    environment: "efa7c243-81aa-4986-b0b1-c89583cbf846",
-    service_entity: "parent-service",
-    version: 4,
-    config: {},
-    state: "up",
-    candidate_attributes: null,
-    active_attributes: {
-      name: "test12345",
-      service_id: "123412",
-      should_deploy_fail: false,
-    },
-    rollback_attributes: null,
-    created_at: "2023-09-19T14:39:30.770002",
-    last_updated: "2023-09-19T14:39:53.389878",
-    callback: [],
-    deleted: false,
-    deployment_progress: null,
-    service_identity_attribute_value: "test12345",
-    referenced_by: [],
-  },
-  interServiceRelations: [
-    {
-      id: "7cd8b669-597a-4341-9b71-07f550b89826",
-      environment: "efa7c243-81aa-4986-b0b1-c89583cbf846",
-      service_entity: "child-service",
-      version: 4,
-      config: {},
-      state: "up",
-      candidate_attributes: null,
-      active_attributes: {
-        name: "child-test",
-        service_id: "123523534623",
-        parent_entity: "085cdf92-0894-4b82-8d46-1dd9552e7ba3",
-        should_deploy_fail: false,
-      },
-      rollback_attributes: null,
-      created_at: "2023-09-19T14:40:08.999123",
-      last_updated: "2023-09-19T14:40:36.178723",
-      callback: [],
-      deleted: false,
-      deployment_progress: null,
-      service_identity_attribute_value: "child-test",
-      referenced_by: [],
-    },
-    {
-      id: "1548332f-86ab-42fe-bd32-4f3adb9e650b",
-      environment: "efa7c243-81aa-4986-b0b1-c89583cbf846",
-      service_entity: "container-service",
-      version: 4,
-      config: {},
-      state: "up",
-      candidate_attributes: null,
-      active_attributes: {
-        name: "test-container1123",
-        service_id: "123412312",
-        child_container: {
-          name: "123124124",
-          parent_entity: "085cdf92-0894-4b82-8d46-1dd9552e7ba3",
-        },
-        should_deploy_fail: false,
-      },
-      rollback_attributes: null,
-      created_at: "2023-09-19T14:39:52.566097",
-      last_updated: "2023-09-19T14:40:14.861344",
-      callback: [],
-      deleted: false,
-      deployment_progress: null,
-      service_identity_attribute_value: "test-container1123",
-      referenced_by: [],
-    },
-  ],
-};
-
-export const mockedInstanceTwo: InstanceWithRelations = {
-  instance: {
-    id: "085cxf92-0894-4ex2-8d46-1gd9552e7ba3",
-    environment: "efa7c243-81aa-4986-b0b1-c89583cbf846",
-    service_entity: "test-service",
-    version: 4,
-    config: {},
-    state: "up",
-    candidate_attributes: null,
-    active_attributes: {
-      name: "test12345",
-      attrOne: "test12345",
-      dictOne: {},
-      attrTwo: "123",
-      attrThree: "456",
-      attrFour: "789",
-      service_id: "012",
-      should_deploy_fail: false,
-      dictTwo: {
-        data: "string",
-      },
-    },
-    rollback_attributes: null,
-    created_at: "2023-09-19T14:39:30.770002",
-    last_updated: "2023-09-19T14:39:53.389878",
-    callback: [],
-    deleted: false,
-    deployment_progress: null,
-    service_identity_attribute_value: "test12345",
-    referenced_by: [],
-  },
-  interServiceRelations: [],
-};
-
-export const mockedInstanceTwoServiceModel: ServiceModel = {
-  attributes: [
-    {
-      name: "name",
-      description: undefined,
-      modifier: "rw",
-      type: "string",
-      default_value: null,
-      default_value_set: false,
-      validation_type: null,
-      validation_parameters: null,
-    },
-    {
-      name: "attrOne",
-      description: undefined,
-      modifier: "rw",
-      type: "string",
-      default_value: null,
-      default_value_set: false,
-      validation_type: null,
-      validation_parameters: null,
-    },
-    {
-      default_value: {},
-      default_value_set: true,
-      description: undefined,
-      modifier: "rw",
-      name: "dictOne",
-      type: "dict",
-      validation_parameters: null,
-      validation_type: null,
-    },
-    {
-      name: "attrTwo",
-      description: undefined,
-      modifier: "rw",
-      type: "string",
-      default_value: null,
-      default_value_set: false,
-      validation_type: null,
-      validation_parameters: null,
-    },
-    {
-      name: "attrThree",
-      description: undefined,
-      modifier: "rw",
-      type: "string",
-      default_value: null,
-      default_value_set: false,
-      validation_type: null,
-      validation_parameters: null,
-    },
-    {
-      name: "attrFour",
-      description: undefined,
-      modifier: "rw",
-      type: "string",
-      default_value: null,
-      default_value_set: false,
-      validation_type: null,
-      validation_parameters: null,
-    },
-    {
-      name: "should_deploy_fail",
-      description: undefined,
-      modifier: "rw+",
-      type: "bool",
-      default_value: false,
-      default_value_set: true,
-      validation_type: null,
-      validation_parameters: null,
-    },
-    {
-      name: "service_id",
-      description: undefined,
-      modifier: "rw",
-      type: "string",
-      default_value: null,
-      default_value_set: false,
-      validation_type: null,
-      validation_parameters: null,
-    },
-    {
-      default_value: {},
-      default_value_set: true,
-      description: undefined,
-      modifier: "rw",
-      name: "dictTwo",
-      type: "dict",
-      validation_parameters: null,
-      validation_type: null,
-    },
-  ],
-  embedded_entities: [],
-  inter_service_relations: [],
-  environment: "fadd0253-fdbc-4822-a046-2c7bb9eaa68a",
-  name: "test-service",
-  description: undefined,
-  lifecycle: {
-    states: [
-      {
-        name: "update_acknowledged_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "deleting",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: true,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "start",
-        label: null,
-        export_resources: false,
-        validate_self: "candidate",
-        validate_others: null,
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_rejected_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_start",
-        label: null,
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "creating",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "terminated",
-        label: null,
-        export_resources: false,
-        validate_self: null,
-        validate_others: null,
-        purge_resources: false,
-        deleted: true,
-        values: {},
-      },
-      {
-        name: "acknowledged",
-        label: null,
-        export_resources: false,
-        validate_self: "candidate",
-        validate_others: null,
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_inprogress",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_start_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "failed",
-        label: "danger",
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "rejected",
-        label: "danger",
-        export_resources: false,
-        validate_self: "candidate",
-        validate_others: null,
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_acknowledged",
-        label: null,
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "up",
-        label: "success",
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_rejected",
-        label: "warning",
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "rollback",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-    ],
-    transfers: [
-      {
-        source: "start",
-        target: "acknowledged",
-        error: "rejected",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: true,
-        config_name: null,
-        description: "initial validation",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "acknowledged",
-        target: "creating",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: false,
-        config_name: null,
-        description: "ack to creating",
-        target_operation: "promote",
-        error_operation: null,
-      },
-      {
-        source: "creating",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "creating done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "up",
-        target: "update_start",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "up to update_start",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_start",
-        target: "update_acknowledged",
-        error: "update_rejected",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: true,
-        config_name: null,
-        description: "update_start to update_ack",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_start_failed",
-        target: "update_acknowledged_failed",
-        error: "update_rejected_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: true,
-        config_name: null,
-        description: "update_start_failed to update_ack_failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_acknowledged",
-        target: "update_inprogress",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: false,
-        config_name: null,
-        description: "update_ack to update_inprogress",
-        target_operation: "promote",
-        error_operation: null,
-      },
-      {
-        source: "update_acknowledged_failed",
-        target: "update_inprogress",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: false,
-        config_name: null,
-        description: "update_ack_failed to update_inprogress_failed",
-        target_operation: "promote",
-        error_operation: null,
-      },
-      {
-        source: "update_inprogress",
-        target: "up",
-        error: "update_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_start",
-        target: "update_start",
-        error: "update_start_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update start failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected",
-        target: "update_rejected",
-        error: "update_rejected_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "upd_reject failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_start_failed",
-        target: "update_start",
-        error: "update_start_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update start recover",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected_failed",
-        target: "update_rejected",
-        error: "update_rejected_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "upd_reject recover",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected",
-        target: "up",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: true,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "abort update",
-        target_operation: "clear candidate",
-        error_operation: null,
-      },
-      {
-        source: "update_rejected_failed",
-        target: "failed",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: true,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "abort update failed",
-        target_operation: "clear candidate",
-        error_operation: null,
-      },
-      {
-        source: "update_failed",
-        target: "rollback",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: true,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rollback in update_failed",
-        target_operation: "rollback",
-        error_operation: null,
-      },
-      {
-        source: "rollback",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rollback done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "up",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "up to failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "failed",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "failed to up",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_failed",
-        target: "up",
-        error: "update_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update_failed to up",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "rejected",
-        target: "start",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rejected to start",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected",
-        target: "update_start",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update_rejected to start_update",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected_failed",
-        target: "update_start_failed",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update_rejected_failed to start_update_failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "deleting",
-        target: "terminated",
-        error: "deleting",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "deletion done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "rejected",
-        target: "terminated",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rejected to terminated",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "up",
-        target: "deleting",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "up to deleting",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "creating",
-        target: "deleting",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "creating to deleting",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_failed",
-        target: "deleting",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update_failed to deleting",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "failed",
-        target: "deleting",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "failed to deleting",
-        target_operation: null,
-        error_operation: null,
-      },
-    ],
-    initial_state: "start",
-    name: "lsm::simple",
-  },
-  config: {},
-  service_identity: "name",
-  service_identity_display_name: undefined,
-  strict_modifier_enforcement: false,
-  instance_summary: {
-    by_state: {
-      up: 1,
-      update_acknowledged_failed: 0,
-      deleting: 0,
-      start: 0,
-      update_rejected_failed: 0,
-      update_start: 0,
-      creating: 0,
-      acknowledged: 0,
-      update_inprogress: 0,
-      update_start_failed: 0,
-      failed: 0,
-      rejected: 0,
-      update_failed: 0,
-      update_acknowledged: 0,
-      update_rejected: 0,
-      rollback: 0,
-    },
-    by_label: {
-      no_label: 0,
-      info: 0,
-      success: 1,
-      danger: 0,
-      warning: 0,
-    },
-    total: 1,
-  },
-  key_attributes: ["dictOne"],
-  owner: null,
-  owned_entities: [],
-};
-
-export const mockedInstanceThree: InstanceWithRelations = {
-  instance: {
-    id: "085cxf92-0894-4ex2-8d46-1gd9552e7ba3",
-    environment: "efa7c243-81aa-4986-b0b1-c89583cbf846",
-    service_entity: "test-service",
-    version: 4,
-    config: {},
-    state: "up",
-    candidate_attributes: null,
-    active_attributes: {
-      name: "test12345",
-      attrOne: "test12345",
-      dictOne: {},
-      attrTwo: "123",
-      attrThree: "456",
-      attrFour: "789",
-      service_id: "012",
-      should_deploy_fail: false,
-      dictTwo: {},
-    },
-    rollback_attributes: null,
-    created_at: "2023-09-19T14:39:30.770002",
-    last_updated: "2023-09-19T14:39:53.389878",
-    callback: [],
-    deleted: false,
-    deployment_progress: null,
-    service_identity_attribute_value: "test12345",
-    referenced_by: [],
-  },
-  interServiceRelations: [],
-};
-
-export const mockedInstanceThreeServiceModel: ServiceModel = {
-  ...mockedInstanceTwoServiceModel,
-  embedded_entities: [
-    {
-      name: "optionalEmbedded",
-      description: "desc",
-      modifier: "rw",
-      lower_limit: 0,
-      upper_limit: 1,
-      attributes: [],
-      embedded_entities: [],
-      inter_service_relations: [],
-      key_attributes: [],
-    },
-  ],
-};
-
-export const containerModel: ServiceModel = {
-  attributes: [
-    {
-      name: "name",
-      description: "",
-      modifier: "rw",
-
-      type: "string",
-      default_value: null,
-      default_value_set: false,
-      validation_type: null,
-      validation_parameters: null,
-    },
-    {
-      name: "service_id",
-      description: "",
-      modifier: "rw",
-
-      type: "string",
-      default_value: null,
-      default_value_set: false,
-      validation_type: null,
-      validation_parameters: null,
-    },
-    {
-      name: "should_deploy_fail",
-      description: "",
-      modifier: "rw+",
-
-      type: "bool",
-      default_value: false,
-      default_value_set: true,
-      validation_type: null,
-      validation_parameters: null,
-    },
-  ],
-  embedded_entities: [
-    {
-      name: "child_container",
-      description: "",
-      modifier: "rw",
-
-      lower_limit: 1,
-      upper_limit: 1,
-      attributes: [
-        {
-          name: "name",
-          description: "",
-          modifier: "rw",
-
-          type: "string",
-          default_value: null,
-          default_value_set: false,
-          validation_type: null,
-          validation_parameters: null,
-        },
-      ],
-      embedded_entities: [],
-      inter_service_relations: [
-        {
-          name: "parent_entity",
-          description: "",
-          modifier: "rw+",
-
-          lower_limit: 1,
-          upper_limit: 1,
-          entity_type: "parent-service",
-        },
-      ],
-
-      key_attributes: ["name"],
-    },
-  ],
-  inter_service_relations: [],
-
-  environment: "89473b11-69dd-4b54-a9ed-eef74e680f80",
-  name: "container-service",
-  description: "",
-  lifecycle: {
-    states: [
-      {
-        name: "update_inprogress",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_start_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "failed",
-        label: "danger",
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "rejected",
-        label: "danger",
-        export_resources: false,
-        validate_self: "candidate",
-        validate_others: null,
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_acknowledged",
-        label: null,
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "up",
-        label: "success",
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_rejected",
-        label: "warning",
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "rollback",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_acknowledged_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "deleting",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: true,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "start",
-        label: null,
-        export_resources: false,
-        validate_self: "candidate",
-        validate_others: null,
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_rejected_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_start",
-        label: null,
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "creating",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "terminated",
-        label: null,
-        export_resources: false,
-        validate_self: null,
-        validate_others: null,
-        purge_resources: false,
-        deleted: true,
-        values: {},
-      },
-      {
-        name: "acknowledged",
-        label: null,
-        export_resources: false,
-        validate_self: "candidate",
-        validate_others: null,
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-    ],
-    transfers: [
-      {
-        source: "start",
-        target: "acknowledged",
-        error: "rejected",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: true,
-        config_name: null,
-        description: "initial validation",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "acknowledged",
-        target: "creating",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: false,
-        config_name: null,
-        description: "ack to creating",
-        target_operation: "promote",
-        error_operation: null,
-      },
-      {
-        source: "creating",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "creating done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "up",
-        target: "update_start",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "up to update_start",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_start",
-        target: "update_acknowledged",
-        error: "update_rejected",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: true,
-        config_name: null,
-        description: "update_start to update_ack",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_start_failed",
-        target: "update_acknowledged_failed",
-        error: "update_rejected_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: true,
-        config_name: null,
-        description: "update_start_failed to update_ack_failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_acknowledged",
-        target: "update_inprogress",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: false,
-        config_name: null,
-        description: "update_ack to update_inprogress",
-        target_operation: "promote",
-        error_operation: null,
-      },
-      {
-        source: "update_acknowledged_failed",
-        target: "update_inprogress",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: false,
-        config_name: null,
-        description: "update_ack_failed to update_inprogress_failed",
-        target_operation: "promote",
-        error_operation: null,
-      },
-      {
-        source: "update_inprogress",
-        target: "up",
-        error: "update_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_start",
-        target: "update_start",
-        error: "update_start_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update start failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected",
-        target: "update_rejected",
-        error: "update_rejected_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "upd_reject failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_start_failed",
-        target: "update_start",
-        error: "update_start_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update start recover",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected_failed",
-        target: "update_rejected",
-        error: "update_rejected_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "upd_reject recover",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected",
-        target: "up",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: true,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "abort update",
-        target_operation: "clear candidate",
-        error_operation: null,
-      },
-      {
-        source: "update_rejected_failed",
-        target: "failed",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: true,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "abort update failed",
-        target_operation: "clear candidate",
-        error_operation: null,
-      },
-      {
-        source: "update_failed",
-        target: "rollback",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: true,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rollback in update_failed",
-        target_operation: "rollback",
-        error_operation: null,
-      },
-      {
-        source: "rollback",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rollback done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "up",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "up to failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "failed",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "failed to up",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_failed",
-        target: "up",
-        error: "update_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update_failed to up",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "rejected",
-        target: "start",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rejected to start",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected",
-        target: "update_start",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update_rejected to start_update",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected_failed",
-        target: "update_start_failed",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update_rejected_failed to start_update_failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "deleting",
-        target: "terminated",
-        error: "deleting",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "deletion done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "rejected",
-        target: "terminated",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rejected to terminated",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "up",
-        target: "deleting",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "up to deleting",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "creating",
-        target: "deleting",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "creating to deleting",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_failed",
-        target: "deleting",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update_failed to deleting",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "failed",
-        target: "deleting",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "failed to deleting",
-        target_operation: null,
-        error_operation: null,
-      },
-    ],
-    initial_state: "start",
-    name: "lsm::simple",
-  },
-  config: {},
-  service_identity: "name",
-  owner: "parent-service",
-  owned_entities: [],
-  strict_modifier_enforcement: true,
-
-  instance_summary: {
-    by_state: {
-      up: 1,
-      update_inprogress: 0,
-      update_start_failed: 0,
-      failed: 0,
-      rejected: 0,
-      update_failed: 0,
-      update_acknowledged: 0,
-      update_rejected: 0,
-      rollback: 0,
-      update_acknowledged_failed: 0,
-      deleting: 0,
-      start: 0,
-      update_rejected_failed: 0,
-      update_start: 0,
-      creating: 0,
-      acknowledged: 0,
-    },
-    by_label: {
-      no_label: 0,
-      info: 0,
-      success: 1,
-      danger: 0,
-      warning: 0,
-    },
-    total: 1,
-  },
-};
-
-export const childModel: ServiceModel = {
-  attributes: [
-    {
-      name: "name",
-      description: "",
-      modifier: "rw",
-
-      type: "string",
-      default_value: null,
-      default_value_set: false,
-      validation_type: null,
-      validation_parameters: null,
-    },
-    {
-      name: "service_id",
-      description: "",
-      modifier: "rw",
-
-      type: "string",
-      default_value: null,
-      default_value_set: false,
-      validation_type: null,
-      validation_parameters: null,
-    },
-    {
-      name: "should_deploy_fail",
-      description: "",
-      modifier: "rw+",
-
-      type: "bool",
-      default_value: false,
-      default_value_set: true,
-      validation_type: null,
-      validation_parameters: null,
-    },
-  ],
-  embedded_entities: [],
-  inter_service_relations: [
-    {
-      name: "parent_entity",
-      description: "",
-      modifier: "rw+",
-
-      lower_limit: 1,
-      upper_limit: 1,
-      entity_type: "parent-service",
-    },
-  ],
-
-  environment: "89473b11-69dd-4b54-a9ed-eef74e680f80",
-  name: "child-service",
-  description: "",
-  lifecycle: {
-    states: [
-      {
-        name: "update_inprogress",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_start_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "failed",
-        label: "danger",
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "rejected",
-        label: "danger",
-        export_resources: false,
-        validate_self: "candidate",
-        validate_others: null,
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_acknowledged",
-        label: null,
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "up",
-        label: "success",
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_rejected",
-        label: "warning",
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "rollback",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_acknowledged_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "deleting",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: true,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "start",
-        label: null,
-        export_resources: false,
-        validate_self: "candidate",
-        validate_others: null,
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_rejected_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_start",
-        label: null,
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "creating",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "terminated",
-        label: null,
-        export_resources: false,
-        validate_self: null,
-        validate_others: null,
-        purge_resources: false,
-        deleted: true,
-        values: {},
-      },
-      {
-        name: "acknowledged",
-        label: null,
-        export_resources: false,
-        validate_self: "candidate",
-        validate_others: null,
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-    ],
-    transfers: [
-      {
-        source: "start",
-        target: "acknowledged",
-        error: "rejected",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: true,
-        config_name: null,
-        description: "initial validation",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "acknowledged",
-        target: "creating",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: false,
-        config_name: null,
-        description: "ack to creating",
-        target_operation: "promote",
-        error_operation: null,
-      },
-      {
-        source: "creating",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "creating done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "up",
-        target: "update_start",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "up to update_start",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_start",
-        target: "update_acknowledged",
-        error: "update_rejected",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: true,
-        config_name: null,
-        description: "update_start to update_ack",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_start_failed",
-        target: "update_acknowledged_failed",
-        error: "update_rejected_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: true,
-        config_name: null,
-        description: "update_start_failed to update_ack_failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_acknowledged",
-        target: "update_inprogress",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: false,
-        config_name: null,
-        description: "update_ack to update_inprogress",
-        target_operation: "promote",
-        error_operation: null,
-      },
-      {
-        source: "update_acknowledged_failed",
-        target: "update_inprogress",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: false,
-        config_name: null,
-        description: "update_ack_failed to update_inprogress_failed",
-        target_operation: "promote",
-        error_operation: null,
-      },
-      {
-        source: "update_inprogress",
-        target: "up",
-        error: "update_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_start",
-        target: "update_start",
-        error: "update_start_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update start failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected",
-        target: "update_rejected",
-        error: "update_rejected_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "upd_reject failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_start_failed",
-        target: "update_start",
-        error: "update_start_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update start recover",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected_failed",
-        target: "update_rejected",
-        error: "update_rejected_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "upd_reject recover",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected",
-        target: "up",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: true,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "abort update",
-        target_operation: "clear candidate",
-        error_operation: null,
-      },
-      {
-        source: "update_rejected_failed",
-        target: "failed",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: true,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "abort update failed",
-        target_operation: "clear candidate",
-        error_operation: null,
-      },
-      {
-        source: "update_failed",
-        target: "rollback",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: true,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rollback in update_failed",
-        target_operation: "rollback",
-        error_operation: null,
-      },
-      {
-        source: "rollback",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rollback done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "up",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "up to failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "failed",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "failed to up",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_failed",
-        target: "up",
-        error: "update_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update_failed to up",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "rejected",
-        target: "start",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rejected to start",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected",
-        target: "update_start",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update_rejected to start_update",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected_failed",
-        target: "update_start_failed",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update_rejected_failed to start_update_failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "deleting",
-        target: "terminated",
-        error: "deleting",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "deletion done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "rejected",
-        target: "terminated",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rejected to terminated",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "up",
-        target: "deleting",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "up to deleting",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "creating",
-        target: "deleting",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "creating to deleting",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_failed",
-        target: "deleting",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update_failed to deleting",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "failed",
-        target: "deleting",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "failed to deleting",
-        target_operation: null,
-        error_operation: null,
-      },
-    ],
-    initial_state: "start",
-    name: "lsm::simple",
-  },
-  config: {},
-  service_identity: "name",
-  owner: "parent-service",
-  owned_entities: [],
-  strict_modifier_enforcement: false,
-
-  instance_summary: {
-    by_state: {
-      up: 1,
-      update_inprogress: 0,
-      update_start_failed: 0,
-      failed: 0,
-      rejected: 0,
-      update_failed: 0,
-      update_acknowledged: 0,
-      update_rejected: 0,
-      rollback: 0,
-      update_acknowledged_failed: 0,
-      deleting: 0,
-      start: 0,
-      update_rejected_failed: 0,
-      update_start: 0,
-      creating: 0,
-      acknowledged: 0,
-    },
-    by_label: {
-      no_label: 0,
-      info: 0,
-      success: 1,
-      danger: 0,
-      warning: 0,
-    },
-    total: 1,
-  },
-};
-
-export const parentModel: ServiceModel = {
-  attributes: [
-    {
-      name: "name",
-      description: undefined,
-      modifier: "rw",
-      attribute_annotations: {},
-      type: "string",
-      default_value: null,
-      default_value_set: false,
-      validation_type: null,
-      validation_parameters: null,
-    },
-    {
-      name: "should_deploy_fail",
-      description: undefined,
-      modifier: "rw+",
-      attribute_annotations: {},
-      type: "bool",
-      default_value: false,
-      default_value_set: true,
-      validation_type: null,
-      validation_parameters: null,
-    },
-    {
-      name: "service_id",
-      description: undefined,
-      modifier: "rw",
-      attribute_annotations: {},
-      type: "string",
-      default_value: null,
-      default_value_set: false,
-      validation_type: null,
-      validation_parameters: null,
-    },
-  ],
-  embedded_entities: [],
-  inter_service_relations: [],
-  environment: "fadd0253-fdbc-4822-a046-2c7bb9eaa68a",
-  name: "parent-service",
-  description: undefined,
-  lifecycle: {
-    states: [
-      {
-        name: "delete_validating_up",
-        label: "success",
-        export_resources: true,
-        validate_self: null,
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_acknowledged_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "start",
-        label: null,
-        export_resources: false,
-        validate_self: "candidate",
-        validate_others: null,
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "delete_validating_update_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: null,
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "acknowledged",
-        label: null,
-        export_resources: false,
-        validate_self: "candidate",
-        validate_others: null,
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "rejected",
-        label: "danger",
-        export_resources: false,
-        validate_self: "candidate",
-        validate_others: null,
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_rejected",
-        label: "warning",
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_rejected_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "creating",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_inprogress",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "failed",
-        label: "danger",
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "up",
-        label: "success",
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "deleting",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: true,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "terminated",
-        label: null,
-        export_resources: false,
-        validate_self: null,
-        validate_others: null,
-        purge_resources: false,
-        deleted: true,
-        values: {},
-      },
-      {
-        name: "rollback",
-        label: null,
-        export_resources: true,
-        validate_self: "active",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_start",
-        label: null,
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "delete_validating_creating",
-        label: "info",
-        export_resources: true,
-        validate_self: null,
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_start_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "delete_validating_failed",
-        label: "warning",
-        export_resources: true,
-        validate_self: null,
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-      {
-        name: "update_acknowledged",
-        label: null,
-        export_resources: true,
-        validate_self: "candidate",
-        validate_others: "active",
-        purge_resources: false,
-        deleted: false,
-        values: {},
-      },
-    ],
-    transfers: [
-      {
-        source: "start",
-        target: "acknowledged",
-        error: "rejected",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: true,
-        config_name: null,
-        description: "initial validation",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "acknowledged",
-        target: "creating",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: false,
-        config_name: null,
-        description: "ack to creating",
-        target_operation: "promote",
-        error_operation: null,
-      },
-      {
-        source: "creating",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "creating done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "up",
-        target: "update_start",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "up to update_start",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_start",
-        target: "update_acknowledged",
-        error: "update_rejected",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: true,
-        config_name: null,
-        description: "update_start to update_ack",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_start_failed",
-        target: "update_acknowledged_failed",
-        error: "update_rejected_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: true,
-        config_name: null,
-        description: "update_start_failed to update_ack_failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_acknowledged",
-        target: "update_inprogress",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: false,
-        config_name: null,
-        description: "update_ack to update_inprogress",
-        target_operation: "promote",
-        error_operation: null,
-      },
-      {
-        source: "update_acknowledged_failed",
-        target: "update_inprogress",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: false,
-        config_name: null,
-        description: "update_ack_failed to update_inprogress_failed",
-        target_operation: "promote",
-        error_operation: null,
-      },
-      {
-        source: "update_inprogress",
-        target: "up",
-        error: "update_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_start",
-        target: "update_start",
-        error: "update_start_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update start failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected",
-        target: "update_rejected",
-        error: "update_rejected_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "upd_reject failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_start_failed",
-        target: "update_start",
-        error: "update_start_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update start recover",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected_failed",
-        target: "update_rejected",
-        error: "update_rejected_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "upd_reject recover",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected",
-        target: "up",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: true,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "abort update",
-        target_operation: "clear candidate",
-        error_operation: null,
-      },
-      {
-        source: "update_rejected_failed",
-        target: "failed",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: true,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "abort update failed",
-        target_operation: "clear candidate",
-        error_operation: null,
-      },
-      {
-        source: "update_failed",
-        target: "rollback",
-        error: null,
-        on_update: false,
-        on_delete: false,
-        api_set_state: true,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rollback in update_failed",
-        target_operation: "rollback",
-        error_operation: null,
-      },
-      {
-        source: "rollback",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rollback done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "up",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "up to failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "failed",
-        target: "up",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "failed to up",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_failed",
-        target: "up",
-        error: "update_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update_failed to up",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "rejected",
-        target: "start",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rejected to start",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected",
-        target: "update_start",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update_rejected to start_update",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_rejected_failed",
-        target: "update_start_failed",
-        error: null,
-        on_update: true,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update_rejected_failed to start_update_failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "deleting",
-        target: "terminated",
-        error: "deleting",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "deletion done",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "rejected",
-        target: "terminated",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "rejected to terminated",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "up",
-        target: "delete_validating_up",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "up to delete_validating_up",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "creating",
-        target: "delete_validating_creating",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "creating to delete_validating_creating",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "failed",
-        target: "delete_validating_failed",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "failed to delete_validating_failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "update_failed",
-        target: "delete_validating_update_failed",
-        error: null,
-        on_update: false,
-        on_delete: true,
-        api_set_state: false,
-        resource_based: false,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "update_failed to delete_validating_update_failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "delete_validating_up",
-        target: "delete_validating_up",
-        error: "delete_validating_up",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "delete_validating_up to delete_validating_up",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "delete_validating_failed",
-        target: "delete_validating_failed",
-        error: "delete_validating_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "delete_validating_failed to delete_validating_failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "delete_validating_creating",
-        target: "delete_validating_creating",
-        error: "delete_validating_creating",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description: "delete_validating_creating to delete_validating_creating",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "delete_validating_update_failed",
-        target: "delete_validating_update_failed",
-        error: "delete_validating_update_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: true,
-        auto: false,
-        validate: false,
-        config_name: null,
-        description:
-          "delete_validating_update_failed to delete_validating_update_failed",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "delete_validating_up",
-        target: "deleting",
-        error: "up",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: true,
-        config_name: null,
-        description: "validating delete",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "delete_validating_creating",
-        target: "deleting",
-        error: "creating",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: true,
-        config_name: null,
-        description: "validating delete",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "delete_validating_failed",
-        target: "deleting",
-        error: "failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: true,
-        config_name: null,
-        description: "validating delete",
-        target_operation: null,
-        error_operation: null,
-      },
-      {
-        source: "delete_validating_update_failed",
-        target: "deleting",
-        error: "update_failed",
-        on_update: false,
-        on_delete: false,
-        api_set_state: false,
-        resource_based: false,
-        auto: true,
-        validate: true,
-        config_name: null,
-        description: "validating delete",
-        target_operation: null,
-        error_operation: null,
-      },
-    ],
-    initial_state: "start",
-    name: "lsm::simple_with_delete_validate",
-  },
-  config: {},
-  service_identity: "name",
-  service_identity_display_name: undefined,
-  strict_modifier_enforcement: true,
-  owner: null,
-  relation_to_owner: null,
-  instance_summary: {
-    by_state: {
-      up: 1,
-      delete_validating_up: 0,
-      update_acknowledged_failed: 0,
-      start: 0,
-      delete_validating_update_failed: 0,
-      acknowledged: 0,
-      rejected: 0,
-      update_rejected: 0,
-      update_rejected_failed: 0,
-      creating: 0,
-      update_inprogress: 0,
-      failed: 0,
-      update_failed: 0,
-      deleting: 0,
-      rollback: 0,
-      update_start: 0,
-      delete_validating_creating: 0,
-      update_start_failed: 0,
-      delete_validating_failed: 0,
-      update_acknowledged: 0,
-    },
-    by_label: {
-      no_label: 0,
-      info: 0,
-      success: 1,
-      danger: 0,
-      warning: 0,
-    },
-    total: 1,
-  },
-  owned_entities: [],
-};
