+++ src/UI/Components/Diagram/Mocks/services.ts
@@ -1,5 +1,4410 @@
 import { ServiceModel } from "@/Core";
 
+export const mockedInstanceTwoServiceModel: ServiceModel = {
+  attributes: [
+    {
+      name: "name",
+      description: undefined,
+      modifier: "rw",
+      type: "string",
+      default_value: null,
+      default_value_set: false,
+      validation_type: null,
+      validation_parameters: null,
+    },
+    {
+      name: "attrOne",
+      description: undefined,
+      modifier: "rw",
+      type: "string",
+      default_value: null,
+      default_value_set: false,
+      validation_type: null,
+      validation_parameters: null,
+    },
+    {
+      default_value: {},
+      default_value_set: true,
+      description: undefined,
+      modifier: "rw",
+      name: "dictOne",
+      type: "dict",
+      validation_parameters: null,
+      validation_type: null,
+    },
+    {
+      name: "attrTwo",
+      description: undefined,
+      modifier: "rw",
+      type: "string",
+      default_value: null,
+      default_value_set: false,
+      validation_type: null,
+      validation_parameters: null,
+    },
+    {
+      name: "attrThree",
+      description: undefined,
+      modifier: "rw",
+      type: "string",
+      default_value: null,
+      default_value_set: false,
+      validation_type: null,
+      validation_parameters: null,
+    },
+    {
+      name: "attrFour",
+      description: undefined,
+      modifier: "rw",
+      type: "string",
+      default_value: null,
+      default_value_set: false,
+      validation_type: null,
+      validation_parameters: null,
+    },
+    {
+      name: "should_deploy_fail",
+      description: undefined,
+      modifier: "rw+",
+      type: "bool",
+      default_value: false,
+      default_value_set: true,
+      validation_type: null,
+      validation_parameters: null,
+    },
+    {
+      name: "service_id",
+      description: undefined,
+      modifier: "rw",
+      type: "string",
+      default_value: null,
+      default_value_set: false,
+      validation_type: null,
+      validation_parameters: null,
+    },
+    {
+      default_value: {},
+      default_value_set: true,
+      description: undefined,
+      modifier: "rw",
+      name: "dictTwo",
+      type: "dict",
+      validation_parameters: null,
+      validation_type: null,
+    },
+  ],
+  embedded_entities: [],
+  inter_service_relations: [],
+  environment: "fadd0253-fdbc-4822-a046-2c7bb9eaa68a",
+  name: "test-service",
+  description: undefined,
+  lifecycle: {
+    states: [
+      {
+        name: "update_acknowledged_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "deleting",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: true,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "start",
+        label: null,
+        export_resources: false,
+        validate_self: "candidate",
+        validate_others: null,
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_rejected_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_start",
+        label: null,
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "creating",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "terminated",
+        label: null,
+        export_resources: false,
+        validate_self: null,
+        validate_others: null,
+        purge_resources: false,
+        deleted: true,
+        values: {},
+      },
+      {
+        name: "acknowledged",
+        label: null,
+        export_resources: false,
+        validate_self: "candidate",
+        validate_others: null,
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_inprogress",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_start_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "failed",
+        label: "danger",
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "rejected",
+        label: "danger",
+        export_resources: false,
+        validate_self: "candidate",
+        validate_others: null,
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_acknowledged",
+        label: null,
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "up",
+        label: "success",
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_rejected",
+        label: "warning",
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "rollback",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+    ],
+    transfers: [
+      {
+        source: "start",
+        target: "acknowledged",
+        error: "rejected",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: true,
+        config_name: null,
+        description: "initial validation",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "acknowledged",
+        target: "creating",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: false,
+        config_name: null,
+        description: "ack to creating",
+        target_operation: "promote",
+        error_operation: null,
+      },
+      {
+        source: "creating",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "creating done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "up",
+        target: "update_start",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "up to update_start",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_start",
+        target: "update_acknowledged",
+        error: "update_rejected",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: true,
+        config_name: null,
+        description: "update_start to update_ack",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_start_failed",
+        target: "update_acknowledged_failed",
+        error: "update_rejected_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: true,
+        config_name: null,
+        description: "update_start_failed to update_ack_failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_acknowledged",
+        target: "update_inprogress",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: false,
+        config_name: null,
+        description: "update_ack to update_inprogress",
+        target_operation: "promote",
+        error_operation: null,
+      },
+      {
+        source: "update_acknowledged_failed",
+        target: "update_inprogress",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: false,
+        config_name: null,
+        description: "update_ack_failed to update_inprogress_failed",
+        target_operation: "promote",
+        error_operation: null,
+      },
+      {
+        source: "update_inprogress",
+        target: "up",
+        error: "update_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_start",
+        target: "update_start",
+        error: "update_start_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update start failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected",
+        target: "update_rejected",
+        error: "update_rejected_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "upd_reject failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_start_failed",
+        target: "update_start",
+        error: "update_start_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update start recover",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected_failed",
+        target: "update_rejected",
+        error: "update_rejected_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "upd_reject recover",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected",
+        target: "up",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: true,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "abort update",
+        target_operation: "clear candidate",
+        error_operation: null,
+      },
+      {
+        source: "update_rejected_failed",
+        target: "failed",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: true,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "abort update failed",
+        target_operation: "clear candidate",
+        error_operation: null,
+      },
+      {
+        source: "update_failed",
+        target: "rollback",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: true,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rollback in update_failed",
+        target_operation: "rollback",
+        error_operation: null,
+      },
+      {
+        source: "rollback",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rollback done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "up",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "up to failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "failed",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "failed to up",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_failed",
+        target: "up",
+        error: "update_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update_failed to up",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "rejected",
+        target: "start",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rejected to start",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected",
+        target: "update_start",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update_rejected to start_update",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected_failed",
+        target: "update_start_failed",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update_rejected_failed to start_update_failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "deleting",
+        target: "terminated",
+        error: "deleting",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "deletion done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "rejected",
+        target: "terminated",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rejected to terminated",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "up",
+        target: "deleting",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "up to deleting",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "creating",
+        target: "deleting",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "creating to deleting",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_failed",
+        target: "deleting",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update_failed to deleting",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "failed",
+        target: "deleting",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "failed to deleting",
+        target_operation: null,
+        error_operation: null,
+      },
+    ],
+    initial_state: "start",
+    name: "lsm::simple",
+  },
+  config: {},
+  service_identity: "name",
+  service_identity_display_name: undefined,
+  strict_modifier_enforcement: false,
+  instance_summary: {
+    by_state: {
+      up: 1,
+      update_acknowledged_failed: 0,
+      deleting: 0,
+      start: 0,
+      update_rejected_failed: 0,
+      update_start: 0,
+      creating: 0,
+      acknowledged: 0,
+      update_inprogress: 0,
+      update_start_failed: 0,
+      failed: 0,
+      rejected: 0,
+      update_failed: 0,
+      update_acknowledged: 0,
+      update_rejected: 0,
+      rollback: 0,
+    },
+    by_label: {
+      no_label: 0,
+      info: 0,
+      success: 1,
+      danger: 0,
+      warning: 0,
+    },
+    total: 1,
+  },
+  key_attributes: ["dictOne"],
+  owner: null,
+  owned_entities: [],
+};
+
+export const testService: ServiceModel = {
+  attributes: [
+    {
+      name: "name",
+      description: "The name of the mpn to deploy.",
+      modifier: "rw",
+      type: "string",
+      default_value: null,
+      default_value_set: false,
+      validation_type: null,
+      validation_parameters: null,
+    },
+    {
+      name: "mgmt_prefix",
+      description:
+        "The network to use for the management range. This needs to be at least a /25\nso that two /26 can be created from it.",
+      modifier: "rw",
+      type: "string",
+      default_value: null,
+      default_value_set: false,
+      validation_type: "ipaddress.IPv4Network",
+      validation_parameters: null,
+    },
+    {
+      name: "mcc",
+      description: "Mobile country code",
+      modifier: "rw",
+      type: "string",
+      default_value: null,
+      default_value_set: false,
+      validation_type: null,
+      validation_parameters: null,
+    },
+    {
+      name: "mnc",
+      description: "Mobile network code",
+      modifier: "rw",
+      type: "string",
+      default_value: null,
+      default_value_set: false,
+      validation_type: null,
+      validation_parameters: null,
+    },
+    {
+      name: "tenant_group",
+      description: "desc",
+      modifier: "rw",
+      type: "string",
+      default_value: null,
+      default_value_set: false,
+      validation_type: null,
+      validation_parameters: null,
+    },
+    {
+      name: "tenant",
+      description: "desc",
+      modifier: "rw",
+      type: "string",
+      default_value: null,
+      default_value_set: false,
+      validation_type: null,
+      validation_parameters: null,
+    },
+    {
+      name: "site",
+      description: "desc",
+      modifier: "rw",
+      type: "string",
+      default_value: null,
+      default_value_set: false,
+      validation_type: null,
+      validation_parameters: null,
+    },
+  ],
+  embedded_entities: [
+    {
+      name: "peerings",
+      description: "desc",
+      modifier: "rw+",
+      lower_limit: 0,
+      upper_limit: undefined,
+      attributes: [
+        {
+          name: "switch",
+          description: "The name/id of the switch to do the peering on",
+          modifier: "rw",
+          type: "string",
+          default_value: null,
+          default_value_set: false,
+          validation_type: "enum",
+          validation_parameters: {
+            names: {
+              "switch-1": "switch-1",
+              "switch-2": "switch-2",
+              "switch-3": "switch-3",
+              "switch-4": "switch-4",
+            },
+            value: "switch_t",
+          },
+        },
+        {
+          name: "port_name",
+          description:
+            "The name of the port on the switch to create the peering on",
+          modifier: "rw",
+          type: "string",
+          default_value: null,
+          default_value_set: false,
+          validation_type: null,
+          validation_parameters: null,
+        },
+        {
+          name: "local_address",
+          description: "The address to set on the interface on the MPN",
+          modifier: "rw",
+          type: "string",
+          default_value: null,
+          default_value_set: false,
+          validation_type: "ipaddress.IPv4Interface",
+          validation_parameters: null,
+        },
+        {
+          name: "peer_address",
+          description:
+            "The address of the peer. When BPG is used, this is also the peering ip",
+          modifier: "rw",
+          type: "string?",
+          default_value: null,
+          default_value_set: true,
+          validation_type: "ipaddress.IPv4Address?",
+          validation_parameters: null,
+        },
+        {
+          name: "vlan",
+          description:
+            "The vlan ID to use for this peering. When the vlan is 0 the traffic is untagged. If the vlan is already used\ninternally in the deployment, a validation error will be given.",
+          modifier: "rw",
+          type: "int",
+          default_value: null,
+          default_value_set: false,
+          validation_type: null,
+          validation_parameters: null,
+        },
+        {
+          name: "peer_asn",
+          description: "The ASN of the peer to setup the BGP session.",
+          modifier: "rw",
+          type: "int?",
+          default_value: null,
+          default_value_set: true,
+          validation_type: null,
+          validation_parameters: null,
+        },
+        {
+          name: "routes",
+          description:
+            "Routes are structured as prefix,gateway For example: 0.0.0.0/0,192.0.2.1",
+          modifier: "rw",
+          type: "string[]",
+          default_value: [] as string[],
+          default_value_set: true,
+          validation_type: "pydantic.constr[]",
+          validation_parameters: {
+            regex:
+              "^\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}\\/\\d{1,2},\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}$",
+          },
+        },
+      ],
+      embedded_entities: [],
+      inter_service_relations: [],
+      key_attributes: ["port_name", "switch", "vlan"],
+    },
+  ],
+  inter_service_relations: [],
+  environment: "ff6dd80a-e18b-4a06-85fc-9a5cf4b45d0b",
+  name: "mpn",
+  description:
+    "A redundant mobile private network deployment for Core and Transport",
+  lifecycle: {
+    states: [
+      {
+        name: "update_inprogress",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_start_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "failed",
+        label: "danger",
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "rejected",
+        label: "danger",
+        export_resources: false,
+        validate_self: "candidate",
+        validate_others: null,
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_acknowledged",
+        label: null,
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "up",
+        label: "success",
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_rejected",
+        label: "warning",
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "rollback",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "start",
+        label: null,
+        export_resources: false,
+        validate_self: "candidate",
+        validate_others: null,
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_acknowledged_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "deleting",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: true,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "creating",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_rejected_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_start",
+        label: null,
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "acknowledged",
+        label: null,
+        export_resources: false,
+        validate_self: "candidate",
+        validate_others: null,
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "terminated",
+        label: null,
+        export_resources: false,
+        validate_self: null,
+        validate_others: null,
+        purge_resources: false,
+        deleted: true,
+        values: {},
+      },
+    ],
+    transfers: [
+      {
+        source: "start",
+        target: "acknowledged",
+        error: "rejected",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: true,
+        config_name: null,
+        description: "initial validation",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "acknowledged",
+        target: "creating",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: false,
+        config_name: null,
+        description: "ack to creating",
+        target_operation: "promote",
+        error_operation: null,
+      },
+      {
+        source: "creating",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "creating done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "up",
+        target: "update_start",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "up to update_start",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_start",
+        target: "update_acknowledged",
+        error: "update_rejected",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: true,
+        config_name: null,
+        description: "update_start to update_ack",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_start_failed",
+        target: "update_acknowledged_failed",
+        error: "update_rejected_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: true,
+        config_name: null,
+        description: "update_start_failed to update_ack_failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_acknowledged",
+        target: "update_inprogress",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: false,
+        config_name: null,
+        description: "update_ack to update_inprogress",
+        target_operation: "promote",
+        error_operation: null,
+      },
+      {
+        source: "update_acknowledged_failed",
+        target: "update_inprogress",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: false,
+        config_name: null,
+        description: "update_ack_failed to update_inprogress_failed",
+        target_operation: "promote",
+        error_operation: null,
+      },
+      {
+        source: "update_inprogress",
+        target: "up",
+        error: "update_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_start",
+        target: "update_start",
+        error: "update_start_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update start failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected",
+        target: "update_rejected",
+        error: "update_rejected_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "upd_reject failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_start_failed",
+        target: "update_start",
+        error: "update_start_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update start recover",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected_failed",
+        target: "update_rejected",
+        error: "update_rejected_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "upd_reject recover",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected",
+        target: "up",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: true,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "abort update",
+        target_operation: "clear candidate",
+        error_operation: null,
+      },
+      {
+        source: "update_rejected_failed",
+        target: "failed",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: true,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "abort update failed",
+        target_operation: "clear candidate",
+        error_operation: null,
+      },
+      {
+        source: "update_failed",
+        target: "rollback",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: true,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rollback in update_failed",
+        target_operation: "rollback",
+        error_operation: null,
+      },
+      {
+        source: "rollback",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rollback done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "up",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "up to failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "failed",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "failed to up",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_failed",
+        target: "up",
+        error: "update_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update_failed to up",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "rejected",
+        target: "start",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rejected to start",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected",
+        target: "update_start",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update_rejected to start_update",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected_failed",
+        target: "update_start_failed",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update_rejected_failed to start_update_failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "deleting",
+        target: "terminated",
+        error: "deleting",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "deletion done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "rejected",
+        target: "terminated",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rejected to terminated",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "up",
+        target: "deleting",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "up to deleting",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "creating",
+        target: "deleting",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "creating to deleting",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_failed",
+        target: "deleting",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update_failed to deleting",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "failed",
+        target: "deleting",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "failed to deleting",
+        target_operation: null,
+        error_operation: null,
+      },
+    ],
+    initial_state: "start",
+    name: "lsm::simple",
+  },
+  config: {},
+  service_identity: "name",
+  service_identity_display_name: "MPN Name",
+  instance_summary: {
+    by_state: {
+      up: 1,
+      update_inprogress: 0,
+      update_start_failed: 0,
+      failed: 0,
+      rejected: 0,
+      update_failed: 0,
+      update_acknowledged: 0,
+      update_rejected: 0,
+      rollback: 0,
+      start: 0,
+      update_acknowledged_failed: 0,
+      deleting: 0,
+      creating: 0,
+      update_rejected_failed: 0,
+      update_start: 0,
+      acknowledged: 0,
+    },
+    by_label: {
+      no_label: 0,
+      info: 0,
+      success: 1,
+      danger: 0,
+      warning: 0,
+    },
+    total: 1,
+  },
+  strict_modifier_enforcement: true,
+  owner: null,
+  owned_entities: [],
+};
+
+export const testApiInstanceModel: ServiceModel = {
+  attributes: [
+    {
+      name: "name",
+      description: "",
+      modifier: "rw",
+      type: "string",
+      default_value: null,
+      default_value_set: false,
+      validation_type: null,
+      validation_parameters: null,
+    },
+    {
+      name: "service_id",
+      description: "",
+      modifier: "rw",
+      type: "string",
+      default_value: null,
+      default_value_set: false,
+      validation_type: null,
+      validation_parameters: null,
+    },
+    {
+      name: "should_deploy_fail",
+      description: "",
+      modifier: "rw+",
+      type: "bool",
+      default_value: false,
+      default_value_set: true,
+      validation_type: null,
+      validation_parameters: null,
+    },
+  ],
+  embedded_entities: [
+    {
+      name: "vlan_assigment_r1",
+      description: "",
+      modifier: "rw",
+      lower_limit: 1,
+      upper_limit: 1,
+      attributes: [
+        {
+          name: "address",
+          description:
+            "The IP-address/netmask to assign to the given VLAN interface.",
+          modifier: "rw+",
+
+          type: "string",
+          default_value: null,
+          default_value_set: false,
+          validation_parameters: null,
+        },
+        {
+          name: "vlan_id",
+          description: "The VLAN ID to assign to the given interface.",
+          modifier: "rw",
+
+          type: "int",
+          default_value: null,
+          default_value_set: false,
+          validation_type: "pydantic.conint",
+          validation_parameters: {
+            ge: 0,
+            lt: 4096,
+          },
+        },
+        {
+          name: "router_ip",
+          description:
+            "The IP address of the srlinux router that should be configured.",
+          modifier: "rw",
+
+          type: "string",
+          default_value: null,
+          default_value_set: false,
+          validation_parameters: null,
+        },
+        {
+          name: "interface_name",
+          description:
+            "The name of the interface that should be connected on the given VLAN.",
+          modifier: "rw",
+
+          type: "string",
+          default_value: null,
+          default_value_set: false,
+          validation_type: null,
+          validation_parameters: null,
+        },
+      ],
+      embedded_entities: [],
+      inter_service_relations: [],
+      key_attributes: ["interface_name", "router_ip", "vlan_id"],
+    },
+    {
+      name: "vlan_assigment_r2",
+      description: "",
+      modifier: "rw",
+      lower_limit: 1,
+      upper_limit: 1,
+      attributes: [
+        {
+          name: "address",
+          description:
+            "The IP-address/netmask to assign to the given VLAN interface.",
+          modifier: "rw+",
+
+          type: "string",
+          default_value: null,
+          default_value_set: false,
+
+          validation_parameters: null,
+        },
+        {
+          name: "vlan_id",
+          description: "The VLAN ID to assign to the given interface.",
+          modifier: "rw",
+          type: "int",
+          default_value: null,
+          default_value_set: false,
+          validation_type: "pydantic.conint",
+          validation_parameters: {
+            ge: 0,
+            lt: 4096,
+          },
+        },
+        {
+          name: "router_ip",
+          description:
+            "The IP address of the srlinux router that should be configured.",
+          modifier: "rw",
+          type: "string",
+          default_value: null,
+          default_value_set: false,
+          validation_parameters: null,
+        },
+        {
+          name: "interface_name",
+          description:
+            "The name of the interface that should be connected on the given VLAN.",
+          modifier: "rw",
+
+          type: "string",
+          default_value: null,
+          default_value_set: false,
+          validation_type: null,
+          validation_parameters: null,
+        },
+      ],
+      embedded_entities: [],
+      inter_service_relations: [],
+      key_attributes: ["interface_name", "router_ip", "vlan_id"],
+    },
+  ],
+  inter_service_relations: [],
+  environment: "89473b11-69dd-4b54-a9ed-eef74e680f80",
+  name: "embedded-entity-service",
+  description: "",
+  lifecycle: {
+    states: [
+      {
+        name: "update_inprogress",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "creating",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "terminated",
+        label: null,
+        export_resources: false,
+        validate_self: null,
+        validate_others: null,
+        purge_resources: false,
+        deleted: true,
+        values: {},
+      },
+      {
+        name: "setting_start",
+        label: null,
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "failed",
+        label: "danger",
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "start",
+        label: null,
+        export_resources: false,
+        validate_self: "candidate",
+        validate_others: null,
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "setting_inprogress",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "up",
+        label: "success",
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_start",
+        label: null,
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "rejected",
+        label: "danger",
+        export_resources: false,
+        validate_self: "candidate",
+        validate_others: null,
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "deleting",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: true,
+        deleted: false,
+        values: {},
+      },
+    ],
+    transfers: [
+      {
+        source: "start",
+        target: "creating",
+        error: "rejected",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: true,
+        config_name: null,
+        description: "initial validation",
+        target_operation: "promote",
+        error_operation: null,
+      },
+      {
+        source: "creating",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "creating done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "up",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "up to failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "failed",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "failed to up",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "rejected",
+        target: "start",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rejected to start",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "deleting",
+        target: "terminated",
+        error: "deleting",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "resources deleted",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "rejected",
+        target: "terminated",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rejected to terminated",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "up",
+        target: "deleting",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "up to deleting",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "creating",
+        target: "deleting",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "creating to deleting",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "failed",
+        target: "deleting",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "failed to deleting",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "up",
+        target: "update_start",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "up to update_start",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_start",
+        target: "update_inprogress",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: false,
+        config_name: null,
+        description: "update_start to update_inprogress",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_inprogress",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "up",
+        target: "setting_start",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: true,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "up to setting_start",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "setting_start",
+        target: "setting_inprogress",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: false,
+        config_name: null,
+        description: "setting_start to setting_inprogress ",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "setting_inprogress",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "setting done",
+        target_operation: null,
+        error_operation: null,
+      },
+    ],
+    initial_state: "start",
+    name: "frontend_model::simple-with-set-state",
+  },
+  config: {},
+  service_identity: "name",
+  service_identity_display_name: undefined,
+  strict_modifier_enforcement: true,
+  instance_summary: {
+    by_state: {
+      update_inprogress: 0,
+      creating: 0,
+      setting_start: 0,
+      failed: 0,
+      start: 0,
+      setting_inprogress: 0,
+      up: 0,
+      update_start: 0,
+      rejected: 0,
+      deleting: 0,
+    },
+    by_label: {
+      no_label: 0,
+      info: 0,
+      success: 0,
+      danger: 0,
+      warning: 0,
+    },
+    total: 0,
+  },
+  owner: null,
+  owned_entities: [],
+};
+
+export const containerModel: ServiceModel = {
+  attributes: [
+    {
+      name: "name",
+      description: "",
+      modifier: "rw",
+
+      type: "string",
+      default_value: null,
+      default_value_set: false,
+      validation_type: null,
+      validation_parameters: null,
+    },
+    {
+      name: "service_id",
+      description: "",
+      modifier: "rw",
+
+      type: "string",
+      default_value: null,
+      default_value_set: false,
+      validation_type: null,
+      validation_parameters: null,
+    },
+    {
+      name: "should_deploy_fail",
+      description: "",
+      modifier: "rw+",
+
+      type: "bool",
+      default_value: false,
+      default_value_set: true,
+      validation_type: null,
+      validation_parameters: null,
+    },
+  ],
+  embedded_entities: [
+    {
+      name: "child_container",
+      description: "",
+      modifier: "rw",
+
+      lower_limit: 1,
+      upper_limit: 1,
+      attributes: [
+        {
+          name: "name",
+          description: "",
+          modifier: "rw",
+
+          type: "string",
+          default_value: null,
+          default_value_set: false,
+          validation_type: null,
+          validation_parameters: null,
+        },
+      ],
+      embedded_entities: [],
+      inter_service_relations: [
+        {
+          name: "parent_entity",
+          description: "",
+          modifier: "rw+",
+
+          lower_limit: 1,
+          upper_limit: 1,
+          entity_type: "parent-service",
+        },
+      ],
+
+      key_attributes: ["name"],
+    },
+  ],
+  inter_service_relations: [],
+
+  environment: "89473b11-69dd-4b54-a9ed-eef74e680f80",
+  name: "container-service",
+  description: "",
+  lifecycle: {
+    states: [
+      {
+        name: "update_inprogress",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_start_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "failed",
+        label: "danger",
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "rejected",
+        label: "danger",
+        export_resources: false,
+        validate_self: "candidate",
+        validate_others: null,
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_acknowledged",
+        label: null,
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "up",
+        label: "success",
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_rejected",
+        label: "warning",
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "rollback",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_acknowledged_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "deleting",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: true,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "start",
+        label: null,
+        export_resources: false,
+        validate_self: "candidate",
+        validate_others: null,
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_rejected_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_start",
+        label: null,
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "creating",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "terminated",
+        label: null,
+        export_resources: false,
+        validate_self: null,
+        validate_others: null,
+        purge_resources: false,
+        deleted: true,
+        values: {},
+      },
+      {
+        name: "acknowledged",
+        label: null,
+        export_resources: false,
+        validate_self: "candidate",
+        validate_others: null,
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+    ],
+    transfers: [
+      {
+        source: "start",
+        target: "acknowledged",
+        error: "rejected",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: true,
+        config_name: null,
+        description: "initial validation",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "acknowledged",
+        target: "creating",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: false,
+        config_name: null,
+        description: "ack to creating",
+        target_operation: "promote",
+        error_operation: null,
+      },
+      {
+        source: "creating",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "creating done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "up",
+        target: "update_start",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "up to update_start",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_start",
+        target: "update_acknowledged",
+        error: "update_rejected",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: true,
+        config_name: null,
+        description: "update_start to update_ack",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_start_failed",
+        target: "update_acknowledged_failed",
+        error: "update_rejected_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: true,
+        config_name: null,
+        description: "update_start_failed to update_ack_failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_acknowledged",
+        target: "update_inprogress",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: false,
+        config_name: null,
+        description: "update_ack to update_inprogress",
+        target_operation: "promote",
+        error_operation: null,
+      },
+      {
+        source: "update_acknowledged_failed",
+        target: "update_inprogress",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: false,
+        config_name: null,
+        description: "update_ack_failed to update_inprogress_failed",
+        target_operation: "promote",
+        error_operation: null,
+      },
+      {
+        source: "update_inprogress",
+        target: "up",
+        error: "update_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_start",
+        target: "update_start",
+        error: "update_start_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update start failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected",
+        target: "update_rejected",
+        error: "update_rejected_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "upd_reject failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_start_failed",
+        target: "update_start",
+        error: "update_start_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update start recover",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected_failed",
+        target: "update_rejected",
+        error: "update_rejected_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "upd_reject recover",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected",
+        target: "up",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: true,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "abort update",
+        target_operation: "clear candidate",
+        error_operation: null,
+      },
+      {
+        source: "update_rejected_failed",
+        target: "failed",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: true,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "abort update failed",
+        target_operation: "clear candidate",
+        error_operation: null,
+      },
+      {
+        source: "update_failed",
+        target: "rollback",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: true,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rollback in update_failed",
+        target_operation: "rollback",
+        error_operation: null,
+      },
+      {
+        source: "rollback",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rollback done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "up",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "up to failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "failed",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "failed to up",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_failed",
+        target: "up",
+        error: "update_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update_failed to up",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "rejected",
+        target: "start",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rejected to start",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected",
+        target: "update_start",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update_rejected to start_update",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected_failed",
+        target: "update_start_failed",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update_rejected_failed to start_update_failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "deleting",
+        target: "terminated",
+        error: "deleting",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "deletion done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "rejected",
+        target: "terminated",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rejected to terminated",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "up",
+        target: "deleting",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "up to deleting",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "creating",
+        target: "deleting",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "creating to deleting",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_failed",
+        target: "deleting",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update_failed to deleting",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "failed",
+        target: "deleting",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "failed to deleting",
+        target_operation: null,
+        error_operation: null,
+      },
+    ],
+    initial_state: "start",
+    name: "lsm::simple",
+  },
+  config: {},
+  service_identity: "name",
+  owner: "parent-service",
+  owned_entities: [],
+  strict_modifier_enforcement: true,
+
+  instance_summary: {
+    by_state: {
+      up: 1,
+      update_inprogress: 0,
+      update_start_failed: 0,
+      failed: 0,
+      rejected: 0,
+      update_failed: 0,
+      update_acknowledged: 0,
+      update_rejected: 0,
+      rollback: 0,
+      update_acknowledged_failed: 0,
+      deleting: 0,
+      start: 0,
+      update_rejected_failed: 0,
+      update_start: 0,
+      creating: 0,
+      acknowledged: 0,
+    },
+    by_label: {
+      no_label: 0,
+      info: 0,
+      success: 1,
+      danger: 0,
+      warning: 0,
+    },
+    total: 1,
+  },
+};
+
+export const childModel: ServiceModel = {
+  attributes: [
+    {
+      name: "name",
+      description: "",
+      modifier: "rw",
+
+      type: "string",
+      default_value: null,
+      default_value_set: false,
+      validation_type: null,
+      validation_parameters: null,
+    },
+    {
+      name: "service_id",
+      description: "",
+      modifier: "rw",
+
+      type: "string",
+      default_value: null,
+      default_value_set: false,
+      validation_type: null,
+      validation_parameters: null,
+    },
+    {
+      name: "should_deploy_fail",
+      description: "",
+      modifier: "rw+",
+
+      type: "bool",
+      default_value: false,
+      default_value_set: true,
+      validation_type: null,
+      validation_parameters: null,
+    },
+  ],
+  embedded_entities: [],
+  inter_service_relations: [
+    {
+      name: "parent_entity",
+      description: "",
+      modifier: "rw+",
+
+      lower_limit: 1,
+      upper_limit: 1,
+      entity_type: "parent-service",
+    },
+  ],
+
+  environment: "89473b11-69dd-4b54-a9ed-eef74e680f80",
+  name: "child-service",
+  description: "",
+  lifecycle: {
+    states: [
+      {
+        name: "update_inprogress",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_start_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "failed",
+        label: "danger",
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "rejected",
+        label: "danger",
+        export_resources: false,
+        validate_self: "candidate",
+        validate_others: null,
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_acknowledged",
+        label: null,
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "up",
+        label: "success",
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_rejected",
+        label: "warning",
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "rollback",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_acknowledged_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "deleting",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: true,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "start",
+        label: null,
+        export_resources: false,
+        validate_self: "candidate",
+        validate_others: null,
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_rejected_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_start",
+        label: null,
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "creating",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "terminated",
+        label: null,
+        export_resources: false,
+        validate_self: null,
+        validate_others: null,
+        purge_resources: false,
+        deleted: true,
+        values: {},
+      },
+      {
+        name: "acknowledged",
+        label: null,
+        export_resources: false,
+        validate_self: "candidate",
+        validate_others: null,
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+    ],
+    transfers: [
+      {
+        source: "start",
+        target: "acknowledged",
+        error: "rejected",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: true,
+        config_name: null,
+        description: "initial validation",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "acknowledged",
+        target: "creating",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: false,
+        config_name: null,
+        description: "ack to creating",
+        target_operation: "promote",
+        error_operation: null,
+      },
+      {
+        source: "creating",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "creating done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "up",
+        target: "update_start",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "up to update_start",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_start",
+        target: "update_acknowledged",
+        error: "update_rejected",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: true,
+        config_name: null,
+        description: "update_start to update_ack",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_start_failed",
+        target: "update_acknowledged_failed",
+        error: "update_rejected_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: true,
+        config_name: null,
+        description: "update_start_failed to update_ack_failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_acknowledged",
+        target: "update_inprogress",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: false,
+        config_name: null,
+        description: "update_ack to update_inprogress",
+        target_operation: "promote",
+        error_operation: null,
+      },
+      {
+        source: "update_acknowledged_failed",
+        target: "update_inprogress",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: false,
+        config_name: null,
+        description: "update_ack_failed to update_inprogress_failed",
+        target_operation: "promote",
+        error_operation: null,
+      },
+      {
+        source: "update_inprogress",
+        target: "up",
+        error: "update_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_start",
+        target: "update_start",
+        error: "update_start_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update start failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected",
+        target: "update_rejected",
+        error: "update_rejected_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "upd_reject failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_start_failed",
+        target: "update_start",
+        error: "update_start_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update start recover",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected_failed",
+        target: "update_rejected",
+        error: "update_rejected_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "upd_reject recover",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected",
+        target: "up",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: true,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "abort update",
+        target_operation: "clear candidate",
+        error_operation: null,
+      },
+      {
+        source: "update_rejected_failed",
+        target: "failed",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: true,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "abort update failed",
+        target_operation: "clear candidate",
+        error_operation: null,
+      },
+      {
+        source: "update_failed",
+        target: "rollback",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: true,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rollback in update_failed",
+        target_operation: "rollback",
+        error_operation: null,
+      },
+      {
+        source: "rollback",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rollback done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "up",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "up to failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "failed",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "failed to up",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_failed",
+        target: "up",
+        error: "update_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update_failed to up",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "rejected",
+        target: "start",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rejected to start",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected",
+        target: "update_start",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update_rejected to start_update",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected_failed",
+        target: "update_start_failed",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update_rejected_failed to start_update_failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "deleting",
+        target: "terminated",
+        error: "deleting",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "deletion done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "rejected",
+        target: "terminated",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rejected to terminated",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "up",
+        target: "deleting",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "up to deleting",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "creating",
+        target: "deleting",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "creating to deleting",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_failed",
+        target: "deleting",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update_failed to deleting",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "failed",
+        target: "deleting",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "failed to deleting",
+        target_operation: null,
+        error_operation: null,
+      },
+    ],
+    initial_state: "start",
+    name: "lsm::simple",
+  },
+  config: {},
+  service_identity: "name",
+  owner: "parent-service",
+  owned_entities: [],
+  strict_modifier_enforcement: false,
+
+  instance_summary: {
+    by_state: {
+      up: 1,
+      update_inprogress: 0,
+      update_start_failed: 0,
+      failed: 0,
+      rejected: 0,
+      update_failed: 0,
+      update_acknowledged: 0,
+      update_rejected: 0,
+      rollback: 0,
+      update_acknowledged_failed: 0,
+      deleting: 0,
+      start: 0,
+      update_rejected_failed: 0,
+      update_start: 0,
+      creating: 0,
+      acknowledged: 0,
+    },
+    by_label: {
+      no_label: 0,
+      info: 0,
+      success: 1,
+      danger: 0,
+      warning: 0,
+    },
+    total: 1,
+  },
+};
+
+export const parentModel: ServiceModel = {
+  attributes: [
+    {
+      name: "name",
+      description: undefined,
+      modifier: "rw",
+      attribute_annotations: {},
+      type: "string",
+      default_value: null,
+      default_value_set: false,
+      validation_type: null,
+      validation_parameters: null,
+    },
+    {
+      name: "should_deploy_fail",
+      description: undefined,
+      modifier: "rw+",
+      attribute_annotations: {},
+      type: "bool",
+      default_value: false,
+      default_value_set: true,
+      validation_type: null,
+      validation_parameters: null,
+    },
+    {
+      name: "service_id",
+      description: undefined,
+      modifier: "rw",
+      attribute_annotations: {},
+      type: "string",
+      default_value: null,
+      default_value_set: false,
+      validation_type: null,
+      validation_parameters: null,
+    },
+  ],
+  embedded_entities: [],
+  inter_service_relations: [],
+  environment: "fadd0253-fdbc-4822-a046-2c7bb9eaa68a",
+  name: "parent-service",
+  description: undefined,
+  lifecycle: {
+    states: [
+      {
+        name: "delete_validating_up",
+        label: "success",
+        export_resources: true,
+        validate_self: null,
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_acknowledged_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "start",
+        label: null,
+        export_resources: false,
+        validate_self: "candidate",
+        validate_others: null,
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "delete_validating_update_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: null,
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "acknowledged",
+        label: null,
+        export_resources: false,
+        validate_self: "candidate",
+        validate_others: null,
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "rejected",
+        label: "danger",
+        export_resources: false,
+        validate_self: "candidate",
+        validate_others: null,
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_rejected",
+        label: "warning",
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_rejected_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "creating",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_inprogress",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "failed",
+        label: "danger",
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "up",
+        label: "success",
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "deleting",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: true,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "terminated",
+        label: null,
+        export_resources: false,
+        validate_self: null,
+        validate_others: null,
+        purge_resources: false,
+        deleted: true,
+        values: {},
+      },
+      {
+        name: "rollback",
+        label: null,
+        export_resources: true,
+        validate_self: "active",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_start",
+        label: null,
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "delete_validating_creating",
+        label: "info",
+        export_resources: true,
+        validate_self: null,
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_start_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "delete_validating_failed",
+        label: "warning",
+        export_resources: true,
+        validate_self: null,
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+      {
+        name: "update_acknowledged",
+        label: null,
+        export_resources: true,
+        validate_self: "candidate",
+        validate_others: "active",
+        purge_resources: false,
+        deleted: false,
+        values: {},
+      },
+    ],
+    transfers: [
+      {
+        source: "start",
+        target: "acknowledged",
+        error: "rejected",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: true,
+        config_name: null,
+        description: "initial validation",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "acknowledged",
+        target: "creating",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: false,
+        config_name: null,
+        description: "ack to creating",
+        target_operation: "promote",
+        error_operation: null,
+      },
+      {
+        source: "creating",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "creating done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "up",
+        target: "update_start",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "up to update_start",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_start",
+        target: "update_acknowledged",
+        error: "update_rejected",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: true,
+        config_name: null,
+        description: "update_start to update_ack",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_start_failed",
+        target: "update_acknowledged_failed",
+        error: "update_rejected_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: true,
+        config_name: null,
+        description: "update_start_failed to update_ack_failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_acknowledged",
+        target: "update_inprogress",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: false,
+        config_name: null,
+        description: "update_ack to update_inprogress",
+        target_operation: "promote",
+        error_operation: null,
+      },
+      {
+        source: "update_acknowledged_failed",
+        target: "update_inprogress",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: false,
+        config_name: null,
+        description: "update_ack_failed to update_inprogress_failed",
+        target_operation: "promote",
+        error_operation: null,
+      },
+      {
+        source: "update_inprogress",
+        target: "up",
+        error: "update_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_start",
+        target: "update_start",
+        error: "update_start_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update start failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected",
+        target: "update_rejected",
+        error: "update_rejected_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "upd_reject failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_start_failed",
+        target: "update_start",
+        error: "update_start_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update start recover",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected_failed",
+        target: "update_rejected",
+        error: "update_rejected_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "upd_reject recover",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected",
+        target: "up",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: true,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "abort update",
+        target_operation: "clear candidate",
+        error_operation: null,
+      },
+      {
+        source: "update_rejected_failed",
+        target: "failed",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: true,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "abort update failed",
+        target_operation: "clear candidate",
+        error_operation: null,
+      },
+      {
+        source: "update_failed",
+        target: "rollback",
+        error: null,
+        on_update: false,
+        on_delete: false,
+        api_set_state: true,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rollback in update_failed",
+        target_operation: "rollback",
+        error_operation: null,
+      },
+      {
+        source: "rollback",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rollback done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "up",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "up to failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "failed",
+        target: "up",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "failed to up",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_failed",
+        target: "up",
+        error: "update_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update_failed to up",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "rejected",
+        target: "start",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rejected to start",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected",
+        target: "update_start",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update_rejected to start_update",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_rejected_failed",
+        target: "update_start_failed",
+        error: null,
+        on_update: true,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update_rejected_failed to start_update_failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "deleting",
+        target: "terminated",
+        error: "deleting",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "deletion done",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "rejected",
+        target: "terminated",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "rejected to terminated",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "up",
+        target: "delete_validating_up",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "up to delete_validating_up",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "creating",
+        target: "delete_validating_creating",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "creating to delete_validating_creating",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "failed",
+        target: "delete_validating_failed",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "failed to delete_validating_failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "update_failed",
+        target: "delete_validating_update_failed",
+        error: null,
+        on_update: false,
+        on_delete: true,
+        api_set_state: false,
+        resource_based: false,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "update_failed to delete_validating_update_failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "delete_validating_up",
+        target: "delete_validating_up",
+        error: "delete_validating_up",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "delete_validating_up to delete_validating_up",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "delete_validating_failed",
+        target: "delete_validating_failed",
+        error: "delete_validating_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "delete_validating_failed to delete_validating_failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "delete_validating_creating",
+        target: "delete_validating_creating",
+        error: "delete_validating_creating",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description: "delete_validating_creating to delete_validating_creating",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "delete_validating_update_failed",
+        target: "delete_validating_update_failed",
+        error: "delete_validating_update_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: true,
+        auto: false,
+        validate: false,
+        config_name: null,
+        description:
+          "delete_validating_update_failed to delete_validating_update_failed",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "delete_validating_up",
+        target: "deleting",
+        error: "up",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: true,
+        config_name: null,
+        description: "validating delete",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "delete_validating_creating",
+        target: "deleting",
+        error: "creating",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: true,
+        config_name: null,
+        description: "validating delete",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "delete_validating_failed",
+        target: "deleting",
+        error: "failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: true,
+        config_name: null,
+        description: "validating delete",
+        target_operation: null,
+        error_operation: null,
+      },
+      {
+        source: "delete_validating_update_failed",
+        target: "deleting",
+        error: "update_failed",
+        on_update: false,
+        on_delete: false,
+        api_set_state: false,
+        resource_based: false,
+        auto: true,
+        validate: true,
+        config_name: null,
+        description: "validating delete",
+        target_operation: null,
+        error_operation: null,
+      },
+    ],
+    initial_state: "start",
+    name: "lsm::simple_with_delete_validate",
+  },
+  config: {},
+  service_identity: "name",
+  service_identity_display_name: undefined,
+  strict_modifier_enforcement: true,
+  owner: null,
+  relation_to_owner: null,
+  instance_summary: {
+    by_state: {
+      up: 1,
+      delete_validating_up: 0,
+      update_acknowledged_failed: 0,
+      start: 0,
+      delete_validating_update_failed: 0,
+      acknowledged: 0,
+      rejected: 0,
+      update_rejected: 0,
+      update_rejected_failed: 0,
+      creating: 0,
+      update_inprogress: 0,
+      failed: 0,
+      update_failed: 0,
+      deleting: 0,
+      rollback: 0,
+      update_start: 0,
+      delete_validating_creating: 0,
+      update_start_failed: 0,
+      delete_validating_failed: 0,
+      update_acknowledged: 0,
+    },
+    by_label: {
+      no_label: 0,
+      info: 0,
+      success: 1,
+      danger: 0,
+      warning: 0,
+    },
+    total: 1,
+  },
+  owned_entities: [],
+};
+
 export const serviceModels: ServiceModel[] = [
   {
     attributes: [
@@ -1245,1434 +5650,8 @@ export const serviceModels: ServiceModel[] = [
     },
     owned_entities: [],
   },
-  {
-    attributes: [
-      {
-        name: "name",
-        description: null,
-        modifier: "rw",
-        attribute_annotations: {},
-        type: "string",
-        default_value: null,
-        default_value_set: false,
-        validation_type: null,
-        validation_parameters: null,
-      },
-      {
-        name: "should_deploy_fail",
-        description: null,
-        modifier: "rw+",
-        attribute_annotations: {},
-        type: "bool",
-        default_value: false,
-        default_value_set: true,
-        validation_type: null,
-        validation_parameters: null,
-      },
-      {
-        name: "service_id",
-        description: null,
-        modifier: "rw",
-        attribute_annotations: {},
-        type: "string",
-        default_value: null,
-        default_value_set: false,
-        validation_type: null,
-        validation_parameters: null,
-      },
-    ],
-    embedded_entities: [],
-    inter_service_relations: [
-      {
-        name: "parent_entity",
-        description: null,
-        modifier: "rw+",
-        attribute_annotations: {},
-        lower_limit: 1,
-        upper_limit: 1,
-        entity_type: "parent-service",
-      },
-    ],
-    entity_annotations: {},
-    environment: "fadd0253-fdbc-4822-a046-2c7bb9eaa68a",
-    name: "child-service",
-    description: null,
-    lifecycle: {
-      states: [
-        {
-          name: "update_acknowledged_failed",
-          label: "warning",
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "deleting",
-          label: null,
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: true,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "start",
-          label: null,
-          export_resources: false,
-          validate_self: "candidate",
-          validate_others: null,
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_rejected_failed",
-          label: "warning",
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_start",
-          label: null,
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "creating",
-          label: null,
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "terminated",
-          label: null,
-          export_resources: false,
-          validate_self: null,
-          validate_others: null,
-          purge_resources: false,
-          deleted: true,
-          values: {},
-        },
-        {
-          name: "acknowledged",
-          label: null,
-          export_resources: false,
-          validate_self: "candidate",
-          validate_others: null,
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_inprogress",
-          label: null,
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_start_failed",
-          label: "warning",
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "failed",
-          label: "danger",
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "rejected",
-          label: "danger",
-          export_resources: false,
-          validate_self: "candidate",
-          validate_others: null,
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_failed",
-          label: "warning",
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_acknowledged",
-          label: null,
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "up",
-          label: "success",
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_rejected",
-          label: "warning",
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "rollback",
-          label: null,
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-      ],
-      transfers: [
-        {
-          source: "start",
-          target: "acknowledged",
-          error: "rejected",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: true,
-          config_name: null,
-          description: "initial validation",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "acknowledged",
-          target: "creating",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: false,
-          config_name: null,
-          description: "ack to creating",
-          target_operation: "promote",
-          error_operation: null,
-        },
-        {
-          source: "creating",
-          target: "up",
-          error: "failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "creating done",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "up",
-          target: "update_start",
-          error: null,
-          on_update: true,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "up to update_start",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_start",
-          target: "update_acknowledged",
-          error: "update_rejected",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: true,
-          config_name: null,
-          description: "update_start to update_ack",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_start_failed",
-          target: "update_acknowledged_failed",
-          error: "update_rejected_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: true,
-          config_name: null,
-          description: "update_start_failed to update_ack_failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_acknowledged",
-          target: "update_inprogress",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: false,
-          config_name: null,
-          description: "update_ack to update_inprogress",
-          target_operation: "promote",
-          error_operation: null,
-        },
-        {
-          source: "update_acknowledged_failed",
-          target: "update_inprogress",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: false,
-          config_name: null,
-          description: "update_ack_failed to update_inprogress_failed",
-          target_operation: "promote",
-          error_operation: null,
-        },
-        {
-          source: "update_inprogress",
-          target: "up",
-          error: "update_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update done",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_start",
-          target: "update_start",
-          error: "update_start_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update start failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_rejected",
-          target: "update_rejected",
-          error: "update_rejected_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "upd_reject failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_start_failed",
-          target: "update_start",
-          error: "update_start_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update start recover",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_rejected_failed",
-          target: "update_rejected",
-          error: "update_rejected_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "upd_reject recover",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_rejected",
-          target: "up",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: true,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "abort update",
-          target_operation: "clear candidate",
-          error_operation: null,
-        },
-        {
-          source: "update_rejected_failed",
-          target: "failed",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: true,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "abort update failed",
-          target_operation: "clear candidate",
-          error_operation: null,
-        },
-        {
-          source: "update_failed",
-          target: "rollback",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: true,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "rollback in update_failed",
-          target_operation: "rollback",
-          error_operation: null,
-        },
-        {
-          source: "rollback",
-          target: "up",
-          error: "failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "rollback done",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "up",
-          target: "up",
-          error: "failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "up to failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "failed",
-          target: "up",
-          error: "failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "failed to up",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_failed",
-          target: "up",
-          error: "update_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update_failed to up",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "rejected",
-          target: "start",
-          error: null,
-          on_update: true,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "rejected to start",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_rejected",
-          target: "update_start",
-          error: null,
-          on_update: true,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update_rejected to start_update",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_rejected_failed",
-          target: "update_start_failed",
-          error: null,
-          on_update: true,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update_rejected_failed to start_update_failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "deleting",
-          target: "terminated",
-          error: "deleting",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "deletion done",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "rejected",
-          target: "terminated",
-          error: null,
-          on_update: false,
-          on_delete: true,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "rejected to terminated",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "up",
-          target: "deleting",
-          error: null,
-          on_update: false,
-          on_delete: true,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "up to deleting",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "creating",
-          target: "deleting",
-          error: null,
-          on_update: false,
-          on_delete: true,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "creating to deleting",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_failed",
-          target: "deleting",
-          error: null,
-          on_update: false,
-          on_delete: true,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update_failed to deleting",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "failed",
-          target: "deleting",
-          error: null,
-          on_update: false,
-          on_delete: true,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "failed to deleting",
-          target_operation: null,
-          error_operation: null,
-        },
-      ],
-      initial_state: "start",
-      name: "lsm::simple",
-    },
-    config: {},
-    service_identity: "name",
-    service_identity_display_name: null,
-    strict_modifier_enforcement: true,
-    owner: null,
-    relation_to_owner: null,
-    instance_summary: {
-      by_state: {
-        up: 1,
-        update_acknowledged_failed: 0,
-        deleting: 0,
-        start: 0,
-        update_rejected_failed: 0,
-        update_start: 0,
-        creating: 0,
-        acknowledged: 0,
-        update_inprogress: 0,
-        update_start_failed: 0,
-        failed: 0,
-        rejected: 0,
-        update_failed: 0,
-        update_acknowledged: 0,
-        update_rejected: 0,
-        rollback: 0,
-      },
-      by_label: {
-        no_label: 0,
-        info: 0,
-        success: 1,
-        danger: 0,
-        warning: 0,
-      },
-      total: 1,
-    },
-    owned_entities: [],
-  },
-  {
-    attributes: [
-      {
-        name: "name",
-        description: null,
-        modifier: "rw",
-        attribute_annotations: {},
-        type: "string",
-        default_value: null,
-        default_value_set: false,
-        validation_type: null,
-        validation_parameters: null,
-      },
-      {
-        name: "should_deploy_fail",
-        description: null,
-        modifier: "rw+",
-        attribute_annotations: {},
-        type: "bool",
-        default_value: false,
-        default_value_set: true,
-        validation_type: null,
-        validation_parameters: null,
-      },
-      {
-        name: "service_id",
-        description: null,
-        modifier: "rw",
-        attribute_annotations: {},
-        type: "string",
-        default_value: null,
-        default_value_set: false,
-        validation_type: null,
-        validation_parameters: null,
-      },
-    ],
-    embedded_entities: [
-      {
-        name: "child_container",
-        description: null,
-        modifier: "rw",
-        attribute_annotations: {},
-        lower_limit: 1,
-        upper_limit: 1,
-        attributes: [
-          {
-            name: "name",
-            description: null,
-            modifier: "rw",
-            attribute_annotations: {},
-            type: "string",
-            default_value: null,
-            default_value_set: false,
-            validation_type: null,
-            validation_parameters: null,
-          },
-        ],
-        embedded_entities: [],
-        inter_service_relations: [
-          {
-            name: "parent_entity",
-            description: null,
-            modifier: "rw+",
-            attribute_annotations: {},
-            lower_limit: 1,
-            upper_limit: 1,
-            entity_type: "parent-service",
-          },
-        ],
-        entity_annotations: {},
-        key_attributes: ["name"],
-      },
-    ],
-    inter_service_relations: [],
-    entity_annotations: {},
-    environment: "fadd0253-fdbc-4822-a046-2c7bb9eaa68a",
-    name: "container-service",
-    description: null,
-    lifecycle: {
-      states: [
-        {
-          name: "update_acknowledged_failed",
-          label: "warning",
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "deleting",
-          label: null,
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: true,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "start",
-          label: null,
-          export_resources: false,
-          validate_self: "candidate",
-          validate_others: null,
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_rejected_failed",
-          label: "warning",
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_start",
-          label: null,
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "creating",
-          label: null,
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "terminated",
-          label: null,
-          export_resources: false,
-          validate_self: null,
-          validate_others: null,
-          purge_resources: false,
-          deleted: true,
-          values: {},
-        },
-        {
-          name: "acknowledged",
-          label: null,
-          export_resources: false,
-          validate_self: "candidate",
-          validate_others: null,
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_inprogress",
-          label: null,
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_start_failed",
-          label: "warning",
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "failed",
-          label: "danger",
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "rejected",
-          label: "danger",
-          export_resources: false,
-          validate_self: "candidate",
-          validate_others: null,
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_failed",
-          label: "warning",
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_acknowledged",
-          label: null,
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "up",
-          label: "success",
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_rejected",
-          label: "warning",
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "rollback",
-          label: null,
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-      ],
-      transfers: [
-        {
-          source: "start",
-          target: "acknowledged",
-          error: "rejected",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: true,
-          config_name: null,
-          description: "initial validation",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "acknowledged",
-          target: "creating",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: false,
-          config_name: null,
-          description: "ack to creating",
-          target_operation: "promote",
-          error_operation: null,
-        },
-        {
-          source: "creating",
-          target: "up",
-          error: "failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "creating done",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "up",
-          target: "update_start",
-          error: null,
-          on_update: true,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "up to update_start",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_start",
-          target: "update_acknowledged",
-          error: "update_rejected",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: true,
-          config_name: null,
-          description: "update_start to update_ack",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_start_failed",
-          target: "update_acknowledged_failed",
-          error: "update_rejected_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: true,
-          config_name: null,
-          description: "update_start_failed to update_ack_failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_acknowledged",
-          target: "update_inprogress",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: false,
-          config_name: null,
-          description: "update_ack to update_inprogress",
-          target_operation: "promote",
-          error_operation: null,
-        },
-        {
-          source: "update_acknowledged_failed",
-          target: "update_inprogress",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: false,
-          config_name: null,
-          description: "update_ack_failed to update_inprogress_failed",
-          target_operation: "promote",
-          error_operation: null,
-        },
-        {
-          source: "update_inprogress",
-          target: "up",
-          error: "update_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update done",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_start",
-          target: "update_start",
-          error: "update_start_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update start failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_rejected",
-          target: "update_rejected",
-          error: "update_rejected_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "upd_reject failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_start_failed",
-          target: "update_start",
-          error: "update_start_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update start recover",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_rejected_failed",
-          target: "update_rejected",
-          error: "update_rejected_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "upd_reject recover",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_rejected",
-          target: "up",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: true,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "abort update",
-          target_operation: "clear candidate",
-          error_operation: null,
-        },
-        {
-          source: "update_rejected_failed",
-          target: "failed",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: true,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "abort update failed",
-          target_operation: "clear candidate",
-          error_operation: null,
-        },
-        {
-          source: "update_failed",
-          target: "rollback",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: true,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "rollback in update_failed",
-          target_operation: "rollback",
-          error_operation: null,
-        },
-        {
-          source: "rollback",
-          target: "up",
-          error: "failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "rollback done",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "up",
-          target: "up",
-          error: "failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "up to failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "failed",
-          target: "up",
-          error: "failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "failed to up",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_failed",
-          target: "up",
-          error: "update_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update_failed to up",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "rejected",
-          target: "start",
-          error: null,
-          on_update: true,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "rejected to start",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_rejected",
-          target: "update_start",
-          error: null,
-          on_update: true,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update_rejected to start_update",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_rejected_failed",
-          target: "update_start_failed",
-          error: null,
-          on_update: true,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update_rejected_failed to start_update_failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "deleting",
-          target: "terminated",
-          error: "deleting",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "deletion done",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "rejected",
-          target: "terminated",
-          error: null,
-          on_update: false,
-          on_delete: true,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "rejected to terminated",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "up",
-          target: "deleting",
-          error: null,
-          on_update: false,
-          on_delete: true,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "up to deleting",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "creating",
-          target: "deleting",
-          error: null,
-          on_update: false,
-          on_delete: true,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "creating to deleting",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_failed",
-          target: "deleting",
-          error: null,
-          on_update: false,
-          on_delete: true,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update_failed to deleting",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "failed",
-          target: "deleting",
-          error: null,
-          on_update: false,
-          on_delete: true,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "failed to deleting",
-          target_operation: null,
-          error_operation: null,
-        },
-      ],
-      initial_state: "start",
-      name: "lsm::simple",
-    },
-    config: {},
-    service_identity: "name",
-    service_identity_display_name: null,
-    strict_modifier_enforcement: true,
-    owner: null,
-    relation_to_owner: null,
-    instance_summary: {
-      by_state: {
-        up: 1,
-        update_acknowledged_failed: 0,
-        deleting: 0,
-        start: 0,
-        update_rejected_failed: 0,
-        update_start: 0,
-        creating: 0,
-        acknowledged: 0,
-        update_inprogress: 0,
-        update_start_failed: 0,
-        failed: 0,
-        rejected: 0,
-        update_failed: 0,
-        update_acknowledged: 0,
-        update_rejected: 0,
-        rollback: 0,
-      },
-      by_label: {
-        no_label: 0,
-        info: 0,
-        success: 1,
-        danger: 0,
-        warning: 0,
-      },
-      total: 1,
-    },
-    owned_entities: [],
-  },
+  containerModel,
+  childModel,
   {
     attributes: [
       {
@@ -3270,1719 +6249,5 @@ export const serviceModels: ServiceModel[] = [
     },
     owned_entities: [],
   },
-  {
-    attributes: [
-      {
-        name: "name",
-        description: null,
-        modifier: "rw",
-        attribute_annotations: {},
-        type: "string",
-        default_value: null,
-        default_value_set: false,
-        validation_type: null,
-        validation_parameters: null,
-      },
-      {
-        name: "should_deploy_fail",
-        description: null,
-        modifier: "rw+",
-        attribute_annotations: {},
-        type: "bool",
-        default_value: false,
-        default_value_set: true,
-        validation_type: null,
-        validation_parameters: null,
-      },
-      {
-        name: "service_id",
-        description: null,
-        modifier: "rw",
-        attribute_annotations: {},
-        type: "string",
-        default_value: null,
-        default_value_set: false,
-        validation_type: null,
-        validation_parameters: null,
-      },
-    ],
-    embedded_entities: [],
-    inter_service_relations: [],
-    entity_annotations: {},
-    environment: "fadd0253-fdbc-4822-a046-2c7bb9eaa68a",
-    name: "parent-service",
-    description: null,
-    lifecycle: {
-      states: [
-        {
-          name: "delete_validating_up",
-          label: "success",
-          export_resources: true,
-          validate_self: null,
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_acknowledged_failed",
-          label: "warning",
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "start",
-          label: null,
-          export_resources: false,
-          validate_self: "candidate",
-          validate_others: null,
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "delete_validating_update_failed",
-          label: "warning",
-          export_resources: true,
-          validate_self: null,
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "acknowledged",
-          label: null,
-          export_resources: false,
-          validate_self: "candidate",
-          validate_others: null,
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "rejected",
-          label: "danger",
-          export_resources: false,
-          validate_self: "candidate",
-          validate_others: null,
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_rejected",
-          label: "warning",
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_rejected_failed",
-          label: "warning",
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "creating",
-          label: null,
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_inprogress",
-          label: null,
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "failed",
-          label: "danger",
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_failed",
-          label: "warning",
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "up",
-          label: "success",
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "deleting",
-          label: null,
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: true,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "terminated",
-          label: null,
-          export_resources: false,
-          validate_self: null,
-          validate_others: null,
-          purge_resources: false,
-          deleted: true,
-          values: {},
-        },
-        {
-          name: "rollback",
-          label: null,
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_start",
-          label: null,
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "delete_validating_creating",
-          label: "info",
-          export_resources: true,
-          validate_self: null,
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_start_failed",
-          label: "warning",
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "delete_validating_failed",
-          label: "warning",
-          export_resources: true,
-          validate_self: null,
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_acknowledged",
-          label: null,
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-      ],
-      transfers: [
-        {
-          source: "start",
-          target: "acknowledged",
-          error: "rejected",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: true,
-          config_name: null,
-          description: "initial validation",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "acknowledged",
-          target: "creating",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: false,
-          config_name: null,
-          description: "ack to creating",
-          target_operation: "promote",
-          error_operation: null,
-        },
-        {
-          source: "creating",
-          target: "up",
-          error: "failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "creating done",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "up",
-          target: "update_start",
-          error: null,
-          on_update: true,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "up to update_start",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_start",
-          target: "update_acknowledged",
-          error: "update_rejected",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: true,
-          config_name: null,
-          description: "update_start to update_ack",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_start_failed",
-          target: "update_acknowledged_failed",
-          error: "update_rejected_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: true,
-          config_name: null,
-          description: "update_start_failed to update_ack_failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_acknowledged",
-          target: "update_inprogress",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: false,
-          config_name: null,
-          description: "update_ack to update_inprogress",
-          target_operation: "promote",
-          error_operation: null,
-        },
-        {
-          source: "update_acknowledged_failed",
-          target: "update_inprogress",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: false,
-          config_name: null,
-          description: "update_ack_failed to update_inprogress_failed",
-          target_operation: "promote",
-          error_operation: null,
-        },
-        {
-          source: "update_inprogress",
-          target: "up",
-          error: "update_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update done",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_start",
-          target: "update_start",
-          error: "update_start_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update start failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_rejected",
-          target: "update_rejected",
-          error: "update_rejected_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "upd_reject failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_start_failed",
-          target: "update_start",
-          error: "update_start_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update start recover",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_rejected_failed",
-          target: "update_rejected",
-          error: "update_rejected_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "upd_reject recover",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_rejected",
-          target: "up",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: true,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "abort update",
-          target_operation: "clear candidate",
-          error_operation: null,
-        },
-        {
-          source: "update_rejected_failed",
-          target: "failed",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: true,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "abort update failed",
-          target_operation: "clear candidate",
-          error_operation: null,
-        },
-        {
-          source: "update_failed",
-          target: "rollback",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: true,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "rollback in update_failed",
-          target_operation: "rollback",
-          error_operation: null,
-        },
-        {
-          source: "rollback",
-          target: "up",
-          error: "failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "rollback done",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "up",
-          target: "up",
-          error: "failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "up to failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "failed",
-          target: "up",
-          error: "failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "failed to up",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_failed",
-          target: "up",
-          error: "update_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update_failed to up",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "rejected",
-          target: "start",
-          error: null,
-          on_update: true,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "rejected to start",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_rejected",
-          target: "update_start",
-          error: null,
-          on_update: true,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update_rejected to start_update",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_rejected_failed",
-          target: "update_start_failed",
-          error: null,
-          on_update: true,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update_rejected_failed to start_update_failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "deleting",
-          target: "terminated",
-          error: "deleting",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "deletion done",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "rejected",
-          target: "terminated",
-          error: null,
-          on_update: false,
-          on_delete: true,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "rejected to terminated",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "up",
-          target: "delete_validating_up",
-          error: null,
-          on_update: false,
-          on_delete: true,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "up to delete_validating_up",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "creating",
-          target: "delete_validating_creating",
-          error: null,
-          on_update: false,
-          on_delete: true,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "creating to delete_validating_creating",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "failed",
-          target: "delete_validating_failed",
-          error: null,
-          on_update: false,
-          on_delete: true,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "failed to delete_validating_failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_failed",
-          target: "delete_validating_update_failed",
-          error: null,
-          on_update: false,
-          on_delete: true,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update_failed to delete_validating_update_failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "delete_validating_up",
-          target: "delete_validating_up",
-          error: "delete_validating_up",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "delete_validating_up to delete_validating_up",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "delete_validating_failed",
-          target: "delete_validating_failed",
-          error: "delete_validating_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "delete_validating_failed to delete_validating_failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "delete_validating_creating",
-          target: "delete_validating_creating",
-          error: "delete_validating_creating",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description:
-            "delete_validating_creating to delete_validating_creating",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "delete_validating_update_failed",
-          target: "delete_validating_update_failed",
-          error: "delete_validating_update_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description:
-            "delete_validating_update_failed to delete_validating_update_failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "delete_validating_up",
-          target: "deleting",
-          error: "up",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: true,
-          config_name: null,
-          description: "validating delete",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "delete_validating_creating",
-          target: "deleting",
-          error: "creating",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: true,
-          config_name: null,
-          description: "validating delete",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "delete_validating_failed",
-          target: "deleting",
-          error: "failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: true,
-          config_name: null,
-          description: "validating delete",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "delete_validating_update_failed",
-          target: "deleting",
-          error: "update_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: true,
-          config_name: null,
-          description: "validating delete",
-          target_operation: null,
-          error_operation: null,
-        },
-      ],
-      initial_state: "start",
-      name: "lsm::simple_with_delete_validate",
-    },
-    config: {},
-    service_identity: "name",
-    service_identity_display_name: null,
-    strict_modifier_enforcement: true,
-    owner: null,
-    relation_to_owner: null,
-    instance_summary: {
-      by_state: {
-        up: 1,
-        delete_validating_up: 0,
-        update_acknowledged_failed: 0,
-        start: 0,
-        delete_validating_update_failed: 0,
-        acknowledged: 0,
-        rejected: 0,
-        update_rejected: 0,
-        update_rejected_failed: 0,
-        creating: 0,
-        update_inprogress: 0,
-        failed: 0,
-        update_failed: 0,
-        deleting: 0,
-        rollback: 0,
-        update_start: 0,
-        delete_validating_creating: 0,
-        update_start_failed: 0,
-        delete_validating_failed: 0,
-        update_acknowledged: 0,
-      },
-      by_label: {
-        no_label: 0,
-        info: 0,
-        success: 1,
-        danger: 0,
-        warning: 0,
-      },
-      total: 1,
-    },
-    owned_entities: [],
-  },
-  {
-    attributes: [
-      {
-        name: "orderId",
-        description: "Unique identifier of the service order.",
-        modifier: "rw",
-        attribute_annotations: {},
-        type: "string",
-        default_value: null,
-        default_value_set: false,
-        validation_type: null,
-        validation_parameters: null,
-      },
-      {
-        name: "name",
-        description:
-          "The name of the bitstream, it should be unique by onu it is configured for.\ni.e. HSIA, TV, etc.",
-        modifier: "rw",
-        attribute_annotations: {},
-        type: "string",
-        default_value: null,
-        default_value_set: false,
-        validation_type: null,
-        validation_parameters: null,
-      },
-      {
-        name: "customerElid",
-        description:
-          "The internal identifier of the customer, in FNT's inventory.",
-        modifier: "rw",
-        attribute_annotations: {},
-        type: "string",
-        default_value: null,
-        default_value_set: false,
-        validation_type: "pydantic.constr",
-        validation_parameters: {
-          regex: "^[A-Z0-9]+$",
-          strict: true,
-        },
-      },
-      {
-        name: "customerLocationElid",
-        description:
-          "The internal identifier of the customer location where the\nont is installed, in FNT's inventory.",
-        modifier: "rw",
-        attribute_annotations: {},
-        type: "string",
-        default_value: null,
-        default_value_set: false,
-        validation_type: "pydantic.constr",
-        validation_parameters: {
-          regex: "^[A-Z0-9]+$",
-          strict: true,
-        },
-      },
-      {
-        name: "servcoReference",
-        description:
-          "The name of the service connectivity provider, that can be found\nin FNT's inventory.",
-        modifier: "rw",
-        attribute_annotations: {},
-        type: "string",
-        default_value: null,
-        default_value_set: false,
-        validation_type: null,
-        validation_parameters: null,
-      },
-      {
-        name: "qos",
-        description: "The quality of service for the bitstream connection.",
-        modifier: "rw+",
-        attribute_annotations: {},
-        type: "string",
-        default_value: null,
-        default_value_set: false,
-        validation_type: "enum",
-        validation_parameters: {
-          names: {
-            Gold: "Gold",
-            Bronze: "Bronze",
-            Silver: "Silver",
-          },
-          value: "qos_t",
-        },
-      },
-      {
-        name: "bandwidth",
-        description: "The bandwidth available to the bistream connection.",
-        modifier: "rw+",
-        attribute_annotations: {},
-        type: "int",
-        default_value: null,
-        default_value_set: false,
-        validation_type: "pydantic.PositiveInt",
-        validation_parameters: null,
-      },
-      {
-        name: "serviceId",
-        description:
-          "Internal identifier for the epipe deployed from the OLT's PE to the Servco' PE.",
-        modifier: "r",
-        attribute_annotations: {},
-        type: "int?",
-        default_value: null,
-        default_value_set: true,
-        validation_type: "pydantic.PositiveInt?",
-        validation_parameters: null,
-      },
-      {
-        name: "oltVlanId",
-        description: "The vlan identifier for the link OLT <> PE",
-        modifier: "r",
-        attribute_annotations: {},
-        type: "int?",
-        default_value: null,
-        default_value_set: true,
-        validation_type: "pydantic.conint?",
-        validation_parameters: {
-          ge: 1,
-          le: 4094,
-        },
-      },
-      {
-        name: "nniVlanId",
-        description: "The vlan identifier for the servco uplink.",
-        modifier: "r",
-        attribute_annotations: {},
-        type: "int?",
-        default_value: null,
-        default_value_set: true,
-        validation_type: "pydantic.conint?",
-        validation_parameters: {
-          ge: 1,
-          le: 4094,
-        },
-      },
-      {
-        name: "ont",
-        description: null,
-        modifier: "r",
-        attribute_annotations: {},
-        type: "dict?",
-        default_value: null,
-        default_value_set: true,
-        validation_type: null,
-        validation_parameters: null,
-      },
-      {
-        name: "olt",
-        description: null,
-        modifier: "r",
-        attribute_annotations: {},
-        type: "dict?",
-        default_value: null,
-        default_value_set: true,
-        validation_type: null,
-        validation_parameters: null,
-      },
-      {
-        name: "olt_pe",
-        description: null,
-        modifier: "r",
-        attribute_annotations: {},
-        type: "dict?",
-        default_value: null,
-        default_value_set: true,
-        validation_type: null,
-        validation_parameters: null,
-      },
-      {
-        name: "olt_uni",
-        description: null,
-        modifier: "r",
-        attribute_annotations: {},
-        type: "dict?",
-        default_value: null,
-        default_value_set: true,
-        validation_type: null,
-        validation_parameters: null,
-      },
-      {
-        name: "nni_uni",
-        description: null,
-        modifier: "r",
-        attribute_annotations: {},
-        type: "dict?",
-        default_value: null,
-        default_value_set: true,
-        validation_type: null,
-        validation_parameters: null,
-      },
-      {
-        name: "planning_protocol",
-        description: null,
-        modifier: "r",
-        attribute_annotations: {},
-        type: "dict?",
-        default_value: null,
-        default_value_set: true,
-        validation_type: null,
-        validation_parameters: null,
-      },
-    ],
-    embedded_entities: [],
-    inter_service_relations: [],
-    entity_annotations: {},
-    environment: "72c8aa68-c1d4-44bd-8914-e509fdadc24e",
-    name: "FtthBitStream",
-    description: "Bit stream connection from a onu interface to a remote uni.",
-    lifecycle: {
-      states: [
-        {
-          name: "update_rejected_failed",
-          label: "warning",
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_start",
-          label: null,
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "creating",
-          label: null,
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "terminated",
-          label: null,
-          export_resources: false,
-          validate_self: null,
-          validate_others: null,
-          purge_resources: false,
-          deleted: true,
-          values: {},
-        },
-        {
-          name: "acknowledged",
-          label: null,
-          export_resources: false,
-          validate_self: "candidate",
-          validate_others: null,
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_inprogress",
-          label: null,
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_start_failed",
-          label: "warning",
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "failed",
-          label: "danger",
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "rejected",
-          label: "danger",
-          export_resources: false,
-          validate_self: "candidate",
-          validate_others: null,
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_failed",
-          label: "warning",
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_acknowledged",
-          label: null,
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "up",
-          label: "success",
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_rejected",
-          label: "warning",
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "rollback",
-          label: null,
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "update_acknowledged_failed",
-          label: "warning",
-          export_resources: true,
-          validate_self: "candidate",
-          validate_others: "active",
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "deleting",
-          label: null,
-          export_resources: true,
-          validate_self: "active",
-          validate_others: "active",
-          purge_resources: true,
-          deleted: false,
-          values: {},
-        },
-        {
-          name: "start",
-          label: null,
-          export_resources: false,
-          validate_self: "candidate",
-          validate_others: null,
-          purge_resources: false,
-          deleted: false,
-          values: {},
-        },
-      ],
-      transfers: [
-        {
-          source: "start",
-          target: "acknowledged",
-          error: "rejected",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: true,
-          config_name: null,
-          description: "initial validation",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "acknowledged",
-          target: "creating",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: false,
-          config_name: null,
-          description: "ack to creating",
-          target_operation: "promote",
-          error_operation: null,
-        },
-        {
-          source: "creating",
-          target: "up",
-          error: "failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "creating done",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "up",
-          target: "update_start",
-          error: null,
-          on_update: true,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "up to update_start",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_start",
-          target: "update_acknowledged",
-          error: "update_rejected",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: true,
-          config_name: null,
-          description: "update_start to update_ack",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_start_failed",
-          target: "update_acknowledged_failed",
-          error: "update_rejected_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: true,
-          config_name: null,
-          description: "update_start_failed to update_ack_failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_acknowledged",
-          target: "update_inprogress",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: false,
-          config_name: null,
-          description: "update_ack to update_inprogress",
-          target_operation: "promote",
-          error_operation: null,
-        },
-        {
-          source: "update_acknowledged_failed",
-          target: "update_inprogress",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: true,
-          validate: false,
-          config_name: null,
-          description: "update_ack_failed to update_inprogress_failed",
-          target_operation: "promote",
-          error_operation: null,
-        },
-        {
-          source: "update_inprogress",
-          target: "up",
-          error: "update_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update done",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_start",
-          target: "update_start",
-          error: "update_start_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update start failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_rejected",
-          target: "update_rejected",
-          error: "update_rejected_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "upd_reject failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_start_failed",
-          target: "update_start",
-          error: "update_start_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update start recover",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_rejected_failed",
-          target: "update_rejected",
-          error: "update_rejected_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "upd_reject recover",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_rejected",
-          target: "up",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: true,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "abort update",
-          target_operation: "clear candidate",
-          error_operation: null,
-        },
-        {
-          source: "update_rejected_failed",
-          target: "failed",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: true,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "abort update failed",
-          target_operation: "clear candidate",
-          error_operation: null,
-        },
-        {
-          source: "update_failed",
-          target: "rollback",
-          error: null,
-          on_update: false,
-          on_delete: false,
-          api_set_state: true,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "rollback in update_failed",
-          target_operation: "rollback",
-          error_operation: null,
-        },
-        {
-          source: "rollback",
-          target: "up",
-          error: "failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "rollback done",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "up",
-          target: "up",
-          error: "failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "up to failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "failed",
-          target: "up",
-          error: "failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "failed to up",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_failed",
-          target: "up",
-          error: "update_failed",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update_failed to up",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "rejected",
-          target: "start",
-          error: null,
-          on_update: true,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "rejected to start",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_rejected",
-          target: "update_start",
-          error: null,
-          on_update: true,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update_rejected to start_update",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_rejected_failed",
-          target: "update_start_failed",
-          error: null,
-          on_update: true,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update_rejected_failed to start_update_failed",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "deleting",
-          target: "terminated",
-          error: "deleting",
-          on_update: false,
-          on_delete: false,
-          api_set_state: false,
-          resource_based: true,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "deletion done",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "rejected",
-          target: "terminated",
-          error: null,
-          on_update: false,
-          on_delete: true,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "rejected to terminated",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "up",
-          target: "deleting",
-          error: null,
-          on_update: false,
-          on_delete: true,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "up to deleting",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "creating",
-          target: "deleting",
-          error: null,
-          on_update: false,
-          on_delete: true,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "creating to deleting",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "update_failed",
-          target: "deleting",
-          error: null,
-          on_update: false,
-          on_delete: true,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "update_failed to deleting",
-          target_operation: null,
-          error_operation: null,
-        },
-        {
-          source: "failed",
-          target: "deleting",
-          error: null,
-          on_update: false,
-          on_delete: true,
-          api_set_state: false,
-          resource_based: false,
-          auto: false,
-          validate: false,
-          config_name: null,
-          description: "failed to deleting",
-          target_operation: null,
-          error_operation: null,
-        },
-      ],
-      initial_state: "start",
-      name: "lsm::simple",
-    },
-    config: {},
-    service_identity: "orderId",
-    service_identity_display_name: "Order",
-    strict_modifier_enforcement: true,
-    owner: null,
-    relation_to_owner: null,
-    instance_summary: {
-      by_state: {
-        up: 1,
-        update_rejected_failed: 0,
-        update_start: 0,
-        creating: 0,
-        acknowledged: 0,
-        update_inprogress: 0,
-        update_start_failed: 0,
-        failed: 0,
-        rejected: 0,
-        update_failed: 0,
-        update_acknowledged: 0,
-        update_rejected: 0,
-        rollback: 0,
-        update_acknowledged_failed: 0,
-        deleting: 0,
-        start: 0,
-      },
-      by_label: {
-        no_label: 0,
-        info: 0,
-        success: 1,
-        danger: 0,
-        warning: 0,
-      },
-      total: 1,
-    },
-    owned_entities: [],
-  },
+  parentModel,
 ];
