+++ src/UI/Components/Diagram/actions/editMode.test.ts
@@ -1,8 +1,11 @@
 import { dia } from "@inmanta/rappid";
 import { EmbeddedEntity, InstanceAttributeModel, ServiceModel } from "@/Core";
 import { InstanceWithRelations } from "@/Data/Managers/V2/GETTERS/GetInstanceWithRelations";
-import { containerModel, mockedInstanceWithRelations } from "../Mocks";
-import { serviceModels } from "../Mocks/services";
+import {
+  containerModel,
+  mockedInstanceWithRelations,
+  serviceModels,
+} from "../Mocks";
 import {
   appendEmbeddedEntity,
   appendInstance,
